Validate a data frame with abstracts to construct vectpr of strings

Validate a data frame with abstracts to construct vectpr of strings

Usage

abstracts2text(abstracts = NULL)

Arguments

abstracts

An object of the class abstracts (a data frame with abstracts) or a string vector with abstracts

Value

A validated character vector. The function check and encode characters in utf8 standard.

Examples

abstracts <- get_abstracts(lacsSample)
abstracts2text(abstracts)
#>   [1] "abstrak x0d berdasarkan hasil observasi wawancara dan data hasil belajar …"
#>   [2] "abstract background many randomized controlled trials fail to reach their…"
#>   [3] "the pathological findings in japanese raccoon dogs with sarcoptic mange i…"
#>   [4] "we investigated the relationship between prevalence and severity of clini…"
#>   [5] "background the current study sought to understand whether trail users rep…"
#>   [6] "1 background dentine hypersensitivity dh is a frequent clinical problem t…"
#>   [7] "serum samples from feral populations of african green monkeys cercopithec…"
#>   [8] "in this paper the elastic behavior of uniform and functionally graded por…"
#>   [9] "covid 19 is a category b type infection but it has created a serious thre…"
#>  [10] "objective to investigate the serum cystatin cysc chemerin and gastrin rel…"
#>  [11] "abstract the repair of dna double strand breaks dsb is mediated via two m…"
#>  [12] "the iberian lynx lynx pardinus is the most threatened wild feline in the …"
#>  [13] "abstract northern peatlands represent one of the largest carbon pools in …"
#>  [14] "Dobrava virus (DOBV) occurs in two different rodent species, Apodemus fla…"
#>  [15] "philip jessup would not be pleased exactly sixty years after he published…"
#>  [16] "course based undergraduate research experiences cures are an effective wa…"
#>  [17] "this chapter begins by looking at the changing nature of business strateg…"
#>  [18] "the accuracy of a general and a specific method of collecting and countin…"
#>  [19] "in this topic the principles of the modulation of the seeding planting do…"
#>  [20] "the increment of displacement produced in 1000 grams of water at 19 5 c w…"
#>  [21] "two hundred sixty five black bears ursus americanus from northcentral ida…"
#>  [22] "wild japanese serows capricornis crispus were found to have parasitic les…"
#>  [23] "purpose to evaluate the national policy on hearing health care pnasa base…"
#>  [24] "serum samples from 58 mountain lions felis concolor in california usa wer…"
#>  [25] "ticks were collected from vegetation and animals at monthly intervals dur…"
#>  [26] "surface defects mainly intergranular cracks igc in as cast billet is the …"
#>  [27] "to conduct a longitudinal serological survey for evidence of brucella spp…"
#>  [28] "the microstructure and dielectric properties of batio3 based ceramics and…"
#>  [29] "the pacific region consists of numerous small island developing states si…"
#>  [30] "abstract silencing in saccharomyces cerevisiae is found at the mating typ…"
#>  [31] "the gill transit circle at the cape observatory has observed zones 40 52 …"
#>  [32] "rappaport hovav and levin 1998 2010 propose manner result complementarity…"
#>  [33] "in this study emulsification chemical crosslinking method was employed th…"
#>  [34] "severe acute respiratory syndrome coronavirus 2 sars cov 2 and the result…"
#>  [35] "roadway safety is an important aspect of all transportation facilities wi…"
#>  [36] "an apparently novel adenovirus was associated with an epizootic of hemorr…"
#>  [37] "contactless measurements are attractive because they do not contaminate t…"
#>  [38] "bartonella vinsonii subsp berkhoffii is a newly recognized pathogen of do…"
#>  [39] "photomicrography and reflectance microphotometry were used to monitor mel…"
#>  [40] "Hand, foot, and mouth disease (HFMD) caused by enteroviruses remains a pu…"
#>  [41] "this paper discusses the experimental and numerical investigations for th…"
#>  [42] "background instrumented wheelchair wheels can be used to study the kinema…"
#>  [43] "over the last few years sportsmen and women have begun to realise the pot…"
#>  [44] "equine protozoal myeloencephalitis is the most important protozoan diseas…"
#>  [45] "model systems related to non covalent minor groove dna complexes with dis…"
#>  [46] "in this paper we present a novel study on the impact of lossy data compre…"
#>  [47] "abstract x0d this research purposed to search empirical evidence danmenga…"
#>  [48] "abstract background the best choice of biological agents after failure to…"
#>  [49] "the abstracts in two languages can be found in the pdf file of the articl…"
#>  [50] "abstract even though the solutions of numerical reservoir simulation are …"
#>  [51] "abstract a reliable and predictable water supply is essential to any deve…"
#>  [52] "objective to determine within the first 10 weeks post onset the most robu…"
#>  [53] "wastewater treatment can consume a large amount of energy to meet dischar…"
#>  [54] "se presenta una experiencia innovadora llevada a cabo en un club de gimna…"
#>  [55] "bloodsmears from 119 yellow baboons papio hamadryas cynocephalus in six g…"
#>  [56] "d2cco has been investigated in the first excited vibrational states of th…"
#>  [57] "abstract background dentine hypersensitivity can impact functional status…"
#>  [58] "abstract to function properly organisms must adjust their physiology beha…"
#>  [59] "In all herpesviruses a block of genes is present which is composed of the…"
#>  [60] "parasites are characteristically aggregated within hosts but identifying …"
#>  [61] "abstract myelodysplastic syndrome mds a hematologic malignancy is associa…"
#>  [62] "abstract fusarium head blight fhb causes significant grain yield and qual…"
#>  [63] "most cases of streptococcus suis infection in humans are caused by seroty…"
#>  [64] "a crucial step in eye organogenesis is the transition of the optic vesicl…"
#>  [65] "a series of micron sized elastic microspheres were optimally prepared thr…"
#>  [66] "the abstracts in two languages can be found in the pdf file of the articl…"
#>  [67] "abstract suppose that a colleague proposed a fantastic experiment to intr…"
#>  [68] "advertising and marketing is a key aspect of the aesthetic industry howev…"
#>  [69] "abstract oftentimes research ideas in the international accounting field …"
#>  [70] "this study is the first of a series in which a main concern will be to es…"
#>  [71] "goats milk has been said to be a suitable alternative to cows milk for pe…"
#>  [72] "studies on the prevalence of gastrointestinal parasites of chicken reared…"
#>  [73] "parelaphostrongylosis has a rapid onset and is lethal in neonatal moose a…"
#>  [74] "this chapter aims to investigate the acceptance of rights in nineteenth c…"
#>  [75] "abstract purpose dry eye in sjogren s ends up with ulcerations infections…"
#>  [76] "giardia duodenalis is prevalent in tropical settings where diverse opport…"
#>  [77] "abstract this chapter begins with a discussion of individual variations i…"
#>  [78] "in recent years radionuclide therapy rt and targeted radionuclide therapy…"
#>  [79] "sera obtained from pampas deer ozotoceros bezoarticus captured in two dif…"
#>  [80] "inadequate oxygen availability at high altitude leads to oxidative stress…"
#>  [81] "To determine if the emergence of hantavirus pulmonary syndrome in the Uni…"
#>  [82] "one possible explanation of the maintenance of many historical foci of sl…"
#>  [83] "human cryptosporidiosis remains underdiagnosed and rapid accurate diagnos…"
#>  [84] "international trade has experienced a ricardian revival in this article w…"
#>  [85] "the beguine movement is a most remarkable movement in the history of the …"
#>  [86] "abstract against the backdrop of sudden shifts in global political and hi…"
#>  [87] "dmd is an inherited genetic disease affecting 1 in 3500 males it results …"
#>  [88] "abstract in this paper a class of discrete sirs epidemic models with dise…"
#>  [89] "leptospirosis is a common worldwide zoonotic infection is an important li…"
#>  [90] "we described the distribution of badger populations in four different are…"
#>  [91] "background in recent years there has been a surge in the development of f…"
#>  [92] "abstract background studies have reported that women with rheumatoid arth…"
#>  [93] "the scholar the administrator and the manager are adrift on a sea of info…"
#>  [94] "the threat of zoonotic transmission of infectious agents at monkey temple…"
#>  [95] "it has been found that a four constant generalized transcendental equatio…"
#>  [96] "o leary et al o leary et al 2013 science 339 662 667 jats ext link xmlns …"
#>  [97] "with the rapid development of china s economy and society the power suppl…"
#>  [98] "the spectral displacement and the damage state medians were assumed as de…"
#>  [99] "Puumala and Tula viruses are hantaviruses found in Europe and are associa…"
#> [100] "changes in populations of several ungulate species in the serengeti mara …"
#> [101] "the protozoon toxoplasma gondii has a worldwide distribution and affects …"
#> [102] "the hospitality industry has grown in size and complexity to service dema…"
#> [103] "ventral visual stream neural responses are dynamic even for static image …"
#> [104] "complete medical examinations were performed on 25 wild golden crowned si…"
#> [105] "feline coronavirus fcov is reported worldwide and known to cause disease …"
#> [106] "two mountain lions puma con color were found dead in the simi hills area …"
#> [107] "the effect of alkali on acrylic fibers has been studied particularly with…"
#> [108] "dye sensitized solar cells are photoelectrochemical devices which are of …"
#> [109] "ten patients with essential hypertension were hospitalized and treated wi…"
#> [110] "the prevalence of antibodies to toxoplasma gondii was investigated in the…"
#> [111] "abstract objective to compare the fecal incontinence status of patients s…"
#> [112] "during the period of december 2004 to january 2005 bacillus anthracis kil…"
#> [113] "there is an increasing attention towards common bean due to its health be…"
#> [114] "abstract the co existence of hemoglobin h hb h disease and heterozygosity…"
#> [115] "this article reports on the use of an innovative reflective diary method …"
#> [116] "heads of hunter killed white tailed deer odocoileus virginianus were obta…"
#> [117] "we propose a novel method to synthesize high resolution images from image…"
#> [118] "a variety of topics within international marketing are reviewed global pr…"
#> [119] "abstracta survey for ticks on livestock with emphasis on ticks parasitizi…"
#> [120] "summarythe performances of falling film evaporators used in the new zeala…"
#> [121] "stimulation of digestive organs by enteric peptides is lost during total …"
#> [122] "bacteroides fragilis is found among the normal intestinal flora and is in…"
#> [123] "the practice of acupuncture as a method of pain relief is becoming increa…"
#> [124] "for the honghe bridge project located in yunnan province southwest china …"
#> [125] "in the familiar immigration birth death process the events of immigration…"
#> [126] "in sweden the prevalence of trichinella infection in domestic pigs has gr…"
#> [127] "scabies is a major threat to the well being of mountain dwelling bovid ho…"
#> [128] "abstract background pancreatic carcinoma pc is a severe disease associate…"
#> [129] "the ancient silk fabrics are important signs of the development of textil…"
#> [130] "abstract we report here the carbon dioxide co2 budget of a 98 6 km2 subar…"
#> [131] "we conducted a study of the population dynamics movement and diseases of …"
#> [132] "this paper draws on marcia and porfeli s identity status models to examin…"
#> [133] "abstract background the aim of this study was to retrospectively examine …"
#> [134] "abstract myelodysplastic syndromes mds are a heterogeneous disorder of th…"
#> [135] "Street rabies viruses are field isolates known to be highly neurotropic. …"
#> [136] "microsporidial infections may be asymptomatic in immunocompetent hosts bu…"
#> [137] "plasmodium knowlesi a malaria parasite originally thought to be restricte…"
#> [138] "abstract the tail probability p of the distribution of a variable is impo…"
#> [139] "the effects of the fungicide captan on growth development and fine struct…"
#> [140] "the delft self grasping hand is an adjustable passive prosthesis operated…"
#> [141] "sarcocystis neurona was isolated from the brain of a juvenile male southe…"
#> [142] "the study of biodiversity morphological and biological peculiarities of p…"
#> [143] "abstract in this study the effect of shearing speed on edge formability w…"
#> [144] "the prevalence of mycobacterium bovis infection in populations of feral p…"
#> [145] "this article uses a course that meets from 10 p m to 1 a m as a context t…"
#> [146] "cutaneous nodular lesions caused by parapoxvirus infections are commonly …"
#> [147] "this chapter studies conjunctivitis conjunctivitis is a nonspecific term …"
#> [148] "the affordable care act aca required that each state set up a health insu…"
#> [149] "replication of plasmid pbr322 in escherichia coli cells normally requires…"
#> [150] "interventions designed to change behavior delivered to groups rather than…"
#> [151] "abstract ocean energy is one of the most important renewable energy sourc…"
#> [152] "the goal of editing the genomes of stem cells to generate model organisms…"
#> [153] "this paper proposes a novel bilateral zero voltage switching zvs bidirect…"
#> [154] "background acetabular fractures are complex orthopaedic injuries and requ…"
#> [155] "according to recent researches the wireless sensor networks wsn which con…"
#> [156] "carbon dioxide co2 storage in deep saline aquifers is a vital option for …"
#> [157] "the effects on plant growth of bacterial fertilizers prepared from azotob…"
#> [158] "abstractin this paper we focus on isotropic and stationary sphere cross t…"
#> [159] "white tailed deer odocoileus virginianus serve to maintain the neospora c…"
#> [160] "in the developed world acute rheumatic fever arf is rare when it does ari…"
#> [161] "the article considers the essence of a person s suspicion of committing a…"
#> [162] "the paper focuses on the experimental investigation of charge trapping an…"
#> [163] "serum samples from 106 eurasian lynx lynx lynx from across sweden found d…"
#> [164] "the island fox is listed as a threatened species in california a serologi…"
#> [165] "the option of decarbonizing urban freight transport using battery electri…"
#> [166] "to effectively and efficiently build and fund the projected larger fleet …"
#> [167] "reliable biomarkers quantifying neurodegeneration and neuroinflammation i…"
#> [168] "abstract background the link between oral diseases and mortality remains …"
#> [169] "with the advent of wireless technology new tools are available that are i…"
#> [170] "in this study estimating the maturing condition in gardens helps to enhan…"
#> [171] "a recent report from this laboratory chitaley k wingard c webb r branam h…"
#> [172] "hafnia based ferroelectric thin film transistors fetfts are regarded as p…"
#> [173] "light verb constructions e g give a kiss to somebody syntactically reflec…"
#> [174] "arthropods exhibit a large variety of sex determination systems both at t…"
#> [175] "abstract introduction the presence of a posterior malleolar pm fragment h…"
#> [176] "describes a project to link an entire town in the usa with a comprehensiv…"
#> [177] "in this paper a sliding mode controller is developed based on the modifie…"
#> [178] "wheat is one of the most important food crops in russia rust diseases lea…"
#> [179] "we have fabricated birnessite type layered manganese oxide bycathodic red…"
#> [180] "grain boundary and free surface tension for pure copper and copper tin al…"
#> [181] "aim the aim of this study was to explore the role of immune modulation th…"
#> [182] "abstract this chapter focses on compensation for livestock when the gover…"
#> [183] "a shorter and less costly alternative to ichikawa s 1 3 transposition pro…"
#> [184] "prevalence and intensity of parasitism by oestrid larvae in the spanish i…"
#> [185] "the immunohistochemical study of chamois rupicapra rupicapra l skin showe…"
#> [186] "as part of a study to determine the cause s of population decline and low…"
#> [187] "nanostructure decorated n type semiconductor interfaces are studied in or…"
#> [188] "quantum metrological standards based on the fundamental physical constant…"
#> [189] "larvae of mesocestoides sp were recovered in tenerife canary islands in 2…"
#> [190] "abstract the objectives of this work are to understand the characteristic…"
#> [191] "annexin vi is a member of a family of ca 2 dependent phospholipid binding…"
#> [192] "abstract background comparatively few studies address the problems relate…"
#> [193] "background zinc phosphate glass ionomer and resin cement are the most com…"
#> [194] "abstract a conductive polypyrrole ppy has been successfully synthesized a…"
#> [195] "the paper presents a new tracking control approach for a class of non min…"
#> [196] "pyriforms and ring forms of babesia odocoilei were detected in thin blood…"
#> [197] "over the past decade the zaire strain of ebola virus zebov has repeatedly…"
#> [198] "the red fox vulpes vulpes population in norway was na\\xefve to sarcoptes…"
#> [199] "this paper examines the clinical and legal outcomes of a consecutive seri…"
#> [200] "an intersystem or vertical handover is a key enabling mechanism for next …"
#> [201] "abstract high strength steel has high strength and poor plasticity when i…"
#> [202] "there are various types of opinion on disparities between private and pub…"
#> [203] "15529 background the literature is scarce regarding the use of interstiti…"
#> [204] "describes the relationship between formal quality standards and total qua…"
#> [205] "a study of ticks associated with wild animals was carried out from septem…"
#> [206] "electropolishing of stainless steel sus has been commercially performed i…"
#> [207] "objectives this study compared the prevalence of cigarette smoking and al…"
#> [208] "food banks are normally owned by non profit organizations which responsib…"
#> [209] "abstract on july 19th 2013 the oecd published an action plan on base eros…"
#> [210] "the nematode parasite angiostrongylus vasorum is a source of increasing c…"
#> [211] "complex regional pain syndrome is a chronic pain syndrome that is often i…"
#> [212] "older people are a major risk group for falls 35 40 of over 65s living at…"
#> [213] "an 18 mo study was conducted in bulloch county georgia to determine tick …"
#> [214] "experimental studies of hot water depressurization show that the pressure…"
#> [215] "necropsy of 340 stray and semi stray dogs canis familiaris and nine red f…"
#> [216] "purpose to compare the changes in visual and ocular parameters in individ…"
#> [217] "within the past decade machine learning algorithms have been proposed as …"
#> [218] "liquid crystal on silicon lcos spatial light modulator slm is the most wi…"
#> [219] "abstract background rna n 6 methyladenosine m 6a plays a critical role in…"
#> [220] "medically unexplained symptoms mus afflict up to 30 of patients who visit…"
#> [221] "students mathematical understanding is still low so it is needed model le…"
#> [222] "five critically endangered iberian lynxes lynx pardinus and 35 other symp…"
#> [223] "observing the importance of stress among employees this study was conduct…"
#> [224] "to what extent can external incentives influence students effort and lear…"
#> [225] "one of the phenomena of the low dose radiation effect is the radio adapta…"
#> [226] "lt em gt abstract lt em gt the mackenzie river is the second longest rive…"
#> [227] "abstract background parkinson s disease pd is the second most prevalent n…"
#> [228] "retusigaster dangerfield austin amp amp whitfield 1999 is a genus of the …"
#> [229] "abstract the genome of the white rot basidiomycete pleurotus ostreatus in…"
#> [230] "abstract there is concern that the rapid expansion of aquaculture may be …"
#> [231] "bronchoalveolar lavage bal is recognized as an important research tool fo…"
#> [232] "abstract class iii hdacs also known as sirtuins couple deacetylation of t…"
#> [233] "this paper presents two vlsi architectures for the recognition of context…"
#> [234] "historically roofing work has been ranked among the highest of all indust…"
#> [235] "the article discusses the late bronze early iron age clay seals mostly st…"
#> [236] "william foote whyte s defense of the methodology used in street corner so…"
#> [237] "pierre klossowski s last major theoretical text living currency 2017 saw …"
#> [238] "abstract abstract 2883 poster board ii 859 background imgn901 hun901 dm1 …"
#> [239] "a cross sectional parasitologic survey of a population of wild sifaka pro…"
#> [240] "the purpose of this article is to describe the development mission and ev…"
#> [241] "summaryintroduction this article is part of the focus theme of methods of…"
#> [242] "the development of modern informational communication technologies has le…"
#> [243] "between 1989 and 1992 7970 wild ungulates comprising 14 different species…"
#> [244] "subject economic outlook post covid significance in recent days the gover…"
#> [245] "lt b gt lt i gt purpose lt i gt lt b gt the study aimed to evaluate the i…"
#> [246] "blood specimens from wild dogs n 301 were obtained from de wildt cheetah …"
#> [247] "sera from 111 bighorn sheep ovis canadensis and 95 mule deer odocoileus h…"
#> [248] "abstractthe hypoxia inducible factor hif proteins are the master regulato…"
#> [249] "pestivirus infection was identified in 16 of 17 chamois during an outbrea…"
#> [250] "tunnel fires are one of the most dangerous catastrophic events that endan…"
#> [251] "as a result of its close collaboration with reproductive health organisat…"
#> [252] "uk and international case materials news updates and commentary this serv…"
#> [253] "lt p gt lt strong gt objective lt strong gt lt span lang en us gt this st…"
#> [254] "abstractthe early developing capacity of human learners to seek out relia…"
#> [255] "a water based binary polyol process where the entangled conformation of p…"
#> [256] "in this article we survey different types of big data techniques such as …"
#> [257] "the grégoires property la piolaine was situated two kilometres to the eas…"
#> [258] "abstrak latar belakang analisis secara nasional dari riskesdas 2013 dan l…"
#> [259] "because of their constant exposure to normal rodent definitive hosts leas…"
#> [260] "michif has two different morphological exponents of plurality the french …"
#> [261] "when a main supply pipe at a european water treatment plant failed large …"
#> [262] "sudden cardiac death scd often falls under the jurisdiction of the medica…"
#> [263] "the relationship between litter size and weight gain of neonatal guinea p…"
#> [264] "from august 1997 to july 2003 we collected 2 103 fecal samples from free …"
#> [265] "investigations were made on possible reservoir hosts of leishmania donova…"
#> [266] "objective angiotensinogen agt is the only substrate of the renin angioten…"
#> [267] "background as cardiac surgery is complicated time consuming and traumatic…"
#> [268] "the aim of this paper is to prove some common fixed point theorems by usi…"
#> [269] "A paramyxovirus was discovered by chance during the primary culture of gr…"
#> [270] "in this paper we prove that 1 2 harmonic quasiconformal mapping is bi lip…"
#> [271] "on the couch house training the id explains how the intertwining of psych…"
#> [272] "individual variability in delay of gratification dg is associated with a …"
#> [273] "the present study was conducted to investigate the effects of maternal pr…"
#> [274] "the florida panther puma concolor coryi is one of the most endangered mam…"
#> [275] "phytophthora foot rot caused by phytophthora parasitica dast can result i…"
#> [276] "development of natural immunostimulants as dietary additive keeps receivi…"
#> [277] "abstract this chapter describes fisher s reintroduction to civilian life …"
#> [278] "parasitism is expected to impact host morbidity or mortality although the…"
#> [279] "a serologic survey of swift fox vulpes velox and kit fox v macrotis from …"
#> [280] "cysts of giardia sp were detected in feces from the rectum of 20 of 74 pi…"
#> [281] "can an occupier invoke the right of self defence against terrorism stemmi…"
#> [282] "Bufavirus (BuV) was first discovered from feces of children with acute di…"
#> [283] "swift foxes vulpes velox have been proposed as potential carriers of flea…"
#> [284] "abstract background imi have high mortality among ltr prevention is criti…"
#> [285] "abstract mn ii complex compounds were synthesized from the reaction of mn…"
#> [286] "the eurasian badger meles meles is considered to be an important wildlife…"
#> [287] "we wished to demonstrate evidence of the presence of california serogroup…"
#> [288] "the lion panthera leo population in the serengeti ecosystem was recently …"
#> [289] "microcontact printing μcp is a low cost technique for rubber stamping tha…"
#> [290] "n myc downstream regulated gene 1 participates in carcinogenesis angiogen…"
#> [291] "abstract meehanite ge is a medium strength iron having high damping capac…"
#> [292] "abstracta review is given on insulators oxides and nitrides which have be…"
#> [293] "abstract while polymer flooding has widely been used as a successful tech…"
#> [294] "in software engineering personality traits have helped to better understa…"
#> [295] "calcium release activated calcium crac channels mediate influx of calcium…"
#> [296] "we examined individual performance on an embedded figures test in two sep…"
#> [297] "a real time pcr assay for the measurement of gamma interferon ifn gamma m…"
#> [298] "winter cooling and persistent mixing for more than a quarter of year nove…"
#> [299] "there have been an increasing number of studies conducted on community pr…"
#> [300] "helminthological examinations of 30 stoat and 31 weasel carcasses were ca…"
#> [301] "with the application of the software pro engineer the 3 d model of automo…"
#> [302] "the aim of this work was to simulate in 1d the spatio temporal evolution …"
#> [303] "impact of deposition parameters microstructure and growth kinetics analys…"
#> [304] "the extent and direction of hybridization between white tailed deer odoco…"
#> [305] "considering the importance of road safety as discussed in the agenda of w…"
#> [306] "several new topical agents have been introduced recently and it can be ex…"
#> [307] "at the end of the cold war two south american rivals built a system of nu…"
#> [308] "purposethe arrival of global manufacturers has brought about new challeng…"
#> [309] "feline retroviruses are rarely reported in lynx species twenty one iberia…"
#> [310] "significance the question of the whether rodent and primate medial fronta…"
#> [311] "the article by harfi hanissian and crawford on the treatment of status as…"
#> [312] "symptoms of neuromuscular dysfunction are quite common in patients with h…"
#> [313] "abstract the honorable the governor of nusa tenggara barat province the m…"
#> [314] "abstract a peaceful and harmonious life is the dream and the hope of ever…"
#> [315] "fusion proteins involving the retinoic acid receptor alpha rar alpha and …"
#> [316] "surveillance and epidemic modeling were used to study chronic wasting dis…"
#> [317] "toxoplasma gondii isolates can be grouped into 3 genetic lineages type i …"
#> [318] "this bibliography is meant to include all books articles and significant …"
#> [319] "several studies have suggested that women may be more resistant to muscle…"
#> [320] "the chapter proposes three ways of integration of the two different world…"
#> [321] "tps7069 background older adults with all have poor survival outcomes alth…"
#> [322] "abstract heat effects in lubricating films are analyzed on the basis of s…"
#> [323] "abstract the plasticity related protein synaptopodin sp has been implicat…"
#> [324] "this study aims to examine the effect of investment information and overc…"
#> [325] "the preparation and technological characterization of nanosphere formulat…"
#> [326] "the demand for low resource devices has increased rapidly due to the adva…"
#> [327] "the exponential decline of great apes over the past 50 years has resulted…"
#> [328] "between august and december 1996 there was an outbreak of a debilitating …"
#> [329] "the history of modern art is often told through aesthetic breakthroughs t…"
#> [330] "an ac driven noiseless underdamped brownian particle moving on an asymmet…"
#> [331] "abstract prenatal and postpartum anxiety may impair maternal functioning …"
#> [332] "my critical reflection on simone tulumello s intervention focuses first o…"
#> [333] "since roro passenger ships have specific characteristics in handling carg…"
#> [334] "concurrently with a survey for echinococcus multilocularis in the red fox…"
#> [335] "abstract background glycomer 631 and lactomer 9 1 are absorbable suture m…"
#> [336] "abstract we investigate color kinematics duality for gauge theory amplitu…"
#> [337] "Murine adenoviruses (MAdV) are supposedly the oldest members of the genus…"
#> [338] "over 80 of the maxicircle and numerous minicircles of trypanosoma brucei …"
#> [339] "in french in some equative et comparative constructions can be found befo…"
#> [340] "male new zealand white rabbits oryctolagus cuniculus were stereotaxically…"
#> [341] "abstract the primary argument of this essay is that the modern western mu…"
#> [342] "foot and mouth disease is a highly contagious acute viral disease that af…"
#> [343] "geographical and temporal trends in reports of rabid raccoons procyon lot…"
#> [344] "abstract background interleukin 10 il 10 is a cytokine whose main biologi…"
#> [345] "abstract mathematical programming techniques are widely used in the deter…"
#> [346] "we describe a rabies epizootic in northwest alaska usa during 1989 and 19…"
#> [347] "this article identifies and comments on published studies of acupuncture …"
#> [348] "this article describes the research and development with the internationa…"
#> [349] "a membrane bioreactor mbr is one of the modifications to the conventional…"
#> [350] "abstract the net present value npv approach is widely applied to assess f…"
#> [351] "the article explores the problem of periodization of mystical anarchism t…"
#> [352] "to determine if 12 moose alces alces from northern alaska with agglutinat…"
#> [353] "Hantaan (HTN) virus, the etiologic agent of clinically severe hemorrhagic…"
#> [354] "vorticity waves are wave like motions occurring in various types of shear…"
#> [355] "lt i gt background and purpose lt i gt there is epidemiological evidence …"
#> [356] "the onset of metabolic acidosis causes an increased transcription of the …"
#> [357] "the current reappraisal of tradition along with an interest in a music th…"
#> [358] "in this study we provide phylogenetic and biogeographic evidence that the…"
#> [359] "the role of wildlife in tuberculosis epidemiology is being widely studied…"
#> [360] "between 1977 and 2005 six large scale failures of municipal solid waste d…"
#> [361] "there is increasing pressure on the humanities to justify their value and…"
#> [362] "graphene is undoubtedly one of the most attractive material for researche…"
#> [363] "fifty two free ranging blackbuck antelope antilope cervicapra from texas …"
#> [364] "Emerging and re-emerging infectious diseases caused by RNA viruses pose a…"
#> [365] "In the Mekong Delta in southern Vietnam, rats are commonly traded in wet …"
#> [366] "purpose financial burden caused by cancer treatment is associated with ma…"
#> [367] "this paper reports on the findings of a study that focused on the use of …"
#> [368] "the aim of the study was through a research of a company linked to the co…"
#> [369] "a total of 395 coyotes canis latrans was examined in the field for lungwo…"
#> [370] "la universidad es un contexto de aprendizaje en el que el alumnado es el …"
#> [371] "abstract background lenalidomide revlimid has been reported to show activ…"
#> [372] "abstractintroductionhepatosplenic gamma delta t cell lymphoma is a rare p…"
#> [373] "injection of the myotoxin notexin was found to induce regeneration in mus…"
#> [374] "amp lt p amp gt it remains unclear how el ni amp amp 241 o amp amp 8211 s…"
#> [375] "currently the problem of non destructive testing of rail lashes or rollin…"
#> [376] "the paper aims to study the model theoretic properties of differentially …"
#> [377] "evidence from molecular systematic studies suggests that many mushroom sp…"
#> [378] "book review of olexiy haran and maksym yakovlyev editors constructing a p…"
#> [379] "abstract background cognitive deficits may be characteristic for only a s…"
#> [380] "more than 300 polymers of 1 vinyl 2 pyrrolidinone vp were synthesized sub…"
#> [381] "abstrakindikator keberhasilan di usaha kecil dan menenggah ukm adalah pen…"
#> [382] "the correlations between coarsening including grain and pore growth and d…"
#> [383] "we determined the prevalence and incidence of hbv and hcv infection in pe…"
#> [384] "holospora acuminata infects micronuclei of paramecium bursaria protozoa c…"
#> [385] "abstract the incidence of pertussis in the united states has steadily inc…"
#> [386] "the authors investigated the prevalence of salmonella spp in 205 wild bir…"
#> [387] "fibrin clot formation is necessary for maintaining the integrity of the v…"
#> [388] "conventionally prepared immune serum globulin frequently produces severe …"
#> [389] "tables and figures arc an integral part of the medium of communication of…"
#> [390] "to determine the seroprevalence and aspects of the epidemiology of canine…"
#> [391] "We present the molecular identification of Apodemus agrarius (striped fie…"
#> [392] "many amphibian species around the world except in asia suffer morbidity a…"
#> [393] "headline israel merger will move labour more centre than left             "
#> [394] "An ecological assessment of reservoir species was conducted in a rural ar…"
#> [395] "sulfur nanoparticle intercalated coal ldh as a robust electrocatalyst for…"
#> [396] "an epizootic suspected to be caused by parapoxviruses occurred in winter …"
#> [397] "a smart city is one that is highly developed innovative environment frien…"
#> [398] "this article provides a brief presentation of the development of some cog…"
#> [399] "manual curation has long been held to be the gold standard for functional…"
#> [400] "from april 1992 to april 1993 fleas were collected from 15 gray foxes uro…"
#> [401] "abstractbackgroundmutations in brca1 and brca2 cause deficiencies in homo…"
#> [402] "abstract aims we report our experience with a cellular porcine dermal non…"
#> [403] "differentiation between rectal and sigmoid carcinomas is a diagnostic cha…"
#> [404] "the objective of this study was to determine the species spectrum of ixod…"
#> [405] "we present an entropy conservation principle applicable to either discret…"
#> [406] "abstractsalivary glands sgs are very important for maintaining the physio…"
#> [407] "in the course of investigating the malignant catarrhal fever mcf subgroup…"
#> [408] "abstract objective to analyze factors associated with the incomplete time…"
#> [409] "a total of 59 faecal samples from ranging papio anubis baboons and anothe…"
#> [410] "in order to study primate lentivirus evolution in the colobinae subfamily…"
#> [411] "serum samples were collected from 116 wolves which were captured in south…"
#> [412] "having incest offenders reunite with their families is a complex process …"
#> [413] "an orb weaving spider s likelihood of survival is influenced by its abili…"
#> [414] "the life cycles and vertical distributions of limnocalanus macrurus and s…"
#> [415] "abstractone of the causal mechanisms underlying neurodevelopmental disord…"
#> [416] "with the variety of medical imaging equipment s application in the medica…"
#> [417] "this study aims to determine the effect of service quality and public rel…"
#> [418] "The transcription profile of chipmunk parvovirus (ChpPV), a tentative mem…"
#> [419] "abstract the concentrations of the blowing agent and silica filler alter …"
#> [420] "the balanced scorecard bsc has received much literary acclaim in the usa …"
#> [421] "this chapter investigates professional development for esol teachers and …"
#> [422] "this paper uses actor network theory ant and aristotelian virtue ethics t…"
#> [423] "abstractinitially unpoled soft pzt was subjected to a proportional coaxia…"
#> [424] "Hantaviruses, family Bunyaviridae, are rodent-borne RNA viruses that have…"
#> [425] "work instruction for 2xyt and 2xytg medium preparation following this rec…"
#> [426] "rat hepatocytes were preincubated for 16 h with hormones or drugs and the…"
#> [427] "egg excretion of s mattheei was studied for about a year from man baboons…"
#> [428] "в статье анализируются источники личного происхождения иностранцев находи…"
#> [429] "we investigated the comparative efficacy and tolerance of two initial sta…"
#> [430] "this chapter addresses the cultural politics history and revival of vietn…"
#> [431] "like other national urban policy documents the state of the cities report…"
#> [432] "purposethe image of the library and information professional has been a s…"
#> [433] "the first finding of a capillariid in the urinary tract of a free ranging…"
#> [434] "seventy two lynx found dead in the swiss alps and the jura mountains swit…"
#> [435] "the thermal stability of the layered modification of the cu0 5zrte2 polyc…"
#> [436] "a strain of plasmodium brasilianum was isolated from an aotus vociferans …"
#> [437] "serum samples n 457 from wolves canis lupus in northern minnesota were co…"
#> [438] "we describe experiments which demonstrate that shortly after infection of…"
#> [439] "abstract background patients with interstitial lung diseases ilds often h…"
#> [440] "abstract the author presents the proposal for a map as a model based on t…"
#> [441] "load tests are a popular way to diagnose the structural condition of brid…"
#> [442] "abstract this chapter explores the role of the industry in ensuring the s…"
#> [443] "after clostridium botulinum type g organisms and toxin were identified in…"
#> [444] "abstractthe goal of this study is to evaluate the effects of anthropogeni…"
#> [445] "in malaysia malay kampung or villages are modernity s significant other i…"
#> [446] "this paper summarises the presentation by richard windle on media researc…"
#> [447] "dye doped ormosil organically modified silicate nanoparticles nps have si…"
#> [448] "natural rubber is one of the commodities that plays an important role as …"
#> [449] "abstract the horizontal winds in the mesosphere and lower thermosphere ml…"
#> [450] "three new sarcocystis species recovered from muscle fibers of the skeleta…"
#> [451] "the transit motion and the rotating motion have highly different effects …"
#> [452] "a 2 yr retrospective survey 1995 1998 of 63 adult raccoons procyon lotor …"
#> [453] "in this article dedicated to the 95th anniversary of the journal state an…"
#> [454] "we investigate in a sample of 53 emerging markets over 1980 2014 whether …"
#> [455] "background tremendous rise in behavioral disorders in child population de…"
#> [456] "abstractintroductionmeat is a nutrient dense food that plays a central ro…"
#> [457] "the disclosure of hiv serostatus is a difficult emotional task creating o…"
#> [458] "Genetic reassortment has been shown to play an important role in the evol…"
#> [459] "this article collects and examines data relating to the authors of englis…"
#> [460] "the inclusion of marijuana with heroin under the harrison narcotic act an…"
#> [461] "the new orleans police department nopd hypertension project was designed …"
#> [462] "in this paper we propose a new one dimensional two segmental nonlinear ma…"
#> [463] "an endemic area of bovine tuberculosis tb mycobacterium bovis currently a…"
#> [464] "lung group 2 innate lymphoid cells ilc2s drive allergic inflammation and …"
#> [465] "although parasites are important regulatory factors in animal populations…"
#> [466] "vertebral artery dissection may be spontaneous or related to some traumat…"
#> [467] "we compared three transport methods for the recovery of pasteurella haemo…"
#> [468] "we protect ourselves from airborne cross infection in the indoor environm…"
#> [469] "from 1979 to 1985 941 feral swine sus scrofa from 53 locations in 15 stat…"
#> [470] "to facilitate ecotourism and research free ranging mountain gorillas of u…"
#> [471] "although immunodetection of hepatitis b core antigen hbcag in the liver h…"
#> [472] "Hepatitis is a major global health concern. However, the etiology of 10-2…"
#> [473] "sarcocysts from the tongue muscle of a european badger meles meles are re…"
#> [474] "a patient centric patient experience is a key to high quality healthcare …"
#> [475] "sixteen european badgers meles meles from three statutory removal operati…"
#> [476] "learning outcomes the learning outcome is that it will help to sustain yo…"
#> [477] "1 bile increases the digestion of fat when given by the mouth in pill for…"
#> [478] "purposethe purpose of this paper is to describe the influence of the tech…"
#> [479] "cruciate ligament reconstruction has progressed dramatically in the last …"
#> [480] "the induction of neurite outgrowth and arborization is critical for devel…"
#> [481] "understanding the habitat use of wildlife species is important for effect…"
#> [482] "the catalytic activity of natural and synthetic zeolites modified by ion …"
#> [483] "summary translation is important for effective international communicatio…"
#> [484] "objective to evaluate the extent and types of health insurance coverage i…"
#> [485] "the mechanism by which amino acids increase the cellular levels of orotic…"
#> [486] "law enforcement is of concern to law firms a law firm can be understood a…"
#> [487] "this introductory chapter provides an overview of george miller s mad max…"
#> [488] "serum samples from 107 cervids were examined for toxoplasma gondii antibo…"
#> [489] "the impact of economic policies on mental health services and with some d…"
#> [490] "recent molecular exploration of the plasmodium species circulating in gre…"
#> [491] "abstract recent developments in reciprocating mud pump technology provide…"
#> [492] "serum samples from wild hanuman langurs presbytis entellus from mysore st…"
#> [493] "forty caspian seals were surveyed seroepidemiologically between 1993 and …"
#> [494] "a finite difference approach with non uniform meshes was presented for si…"
#> [495] "abstract the thermoacoustic travelling wave device system has been design…"
#> [496] "abstract background regular monitoring of anti malarial drug efficacy is …"
#> [497] "amp lt p amp gt the standard for weather radar nowcasting in the central …"
#> [498] "two hundred and two baylisascaris procyonis were collected from 23 70 of …"
#> [499] "leptospirosis is one of the most widespread bacterial zoonoses in the wor…"
#> [500] "abstractfor every regular cardinal α we construct a cofibrantly generated…"
#> [501] "abstract inducible hematopoietic stem progenitor cell lines represent a m…"
#> [502] "abstract objectives increased metabolic demand and inadequate oral intake…"
#> [503] "the aim of the study was to determine the influence of breed on litter si…"
#> [504] "this research is descriptive research which aims to identify the most use…"
#> [505] "in total 1502 faecal samples were collected from a population of european…"
#> [506] "between 15 november and 13 december 1993 three cases of rabies of probabl…"
#> [507] "abstract with the booming development of china s building industry the re…"
#> [508] "in the modern political world information and technological principles ar…"
#> [509] "the gas transport system on the norwegian continental shelf ncs consists …"
#> [510] "12007 background b inhibits proteasome function and may be synergistic in…"
#> [511] "this article discusses the basic tools of lean production the errors made…"
#> [512] "to assess human exposure to simian immunodeficiency virus siv in west cen…"
#> [513] "we show that gribov diffusion of the partons in the impact parameter plan…"
#> [514] "in anesthetized dogs subarachnoid hemorrhage sah was induced by the mecha…"
#> [515] "summarya seven month old cat was referred after having been missing for o…"
#> [516] "An outbreak of hemorrhagic fever with renal syndrome (HFRS) occurred amon…"
#> [517] "in this paper we present a frequency sweeping excitation and spectrogram …"
#> [518] "this symposium consists of four interrelated presentations and a video ab…"
#> [519] "an adenovirus associated with systemic and localized vascular damage was …"
#> [520] "the concentration of ammonium nitrogen nh4 n frequently exceeds that of n…"
#> [521] "Rodent adenoviruses are important models for human disease. In contrast t…"
#> [522] "a noncytopathic type 1a bovine viral diarrhea virus bvdv was isolated fro…"
#> [523] "abstractantibiotic resistant escherichia coli is one of the most serious …"
#> [524] "we report on the examination of 18 elk alces alces 16 red deer cervus ela…"
#> [525] "background although family technical support seems intuitive there is ver…"
#> [526] "the radioisotopes laboratory was established in january 1972 in order to …"
#> [527] "the conclusion underlines that all the fields analyzed in the book share …"
#> [528] "this transdisciplinary review of research about international cooperation…"
#> [529] "warfare on land occupation of enemy territory judicial functions of the o…"
#> [530] "cryptosporidiosis is a rapidly emerging disease in the tropics this is th…"
#> [531] "background the toxicities of pharmaceuticals for chronic arthritis and os…"
#> [532] "the prevalence of newly diagnosed type 2 diabetes t2d among american adul…"
#> [533] "parasites are linked with their host in a trophic interaction with implic…"
#> [534] "the introduction of competency based medical education curriculum in indi…"
#> [535] "not long ago the university installation i manage was asked by one of our…"
#> [536] "this article explores the supervision of master s students undertaking qu…"
#> [537] "pendrin is an anion exchanger expressed in type b intercalated cells of t…"
#> [538] "the mammalian target of rapamycin mtor signaling pathway which regulates …"
#> [539] "a table as defined by webster s new collegiate dictionary is any collecti…"
#> [540] "recently the mobile wireless communication has seen explosive growth in d…"
#> [541] "the corporate culture of an organization plays an important role in the p…"
#> [542] "abstract a rare cause of mid cavity left ventricular obstruction can be a…"
#> [543] "early events of dna amplification which occur during perturbed replicatio…"
#> [544] "neonothopanus gardneri n gardneri is a species of bioluminescent fungus b…"
#> [545] "objectives the aim of this paper is to examine the recent decision of the…"
#> [546] "four species of ticks found to infest 1391 american mink mustela vison in…"
#> [547] "for use in cviji purification from il 3a virus infected nc64a chlorella   "
#> [548] "past research suggested that when individuals feel that it is their free …"
#> [549] "a total of 559 fleas representing four species pulex irritans ctenocephal…"
#> [550] "the paper presents a typology of professionals who deal with liquidity ma…"
#> [551] "objective to evaluate performance of the first round of hpv based screeni…"
#> [552] "wild canids are reservoir hosts for leishmania infantum and trypanosoma c…"
#> [553] "abstract background positive behaviour and user s collaboration improved …"
#> [554] "purpose the purpose of this study was to examine the reliability of sympt…"
#> [555] "purposethe purpose of this paper is to present and validate a multidimens…"
#> [556] "Picobirnaviruses (PBVs) have recently been classified into the Picobirnav…"
#> [557] "isolates of rhizobium meliloti from indigenous populations at two sites w…"
#> [558] "abstractbroadening the diversity of cultivated soybean glycine max l merr…"
#> [559] "hydrocharis morsus ranae l is a eurasian aquatic plant introduced to nort…"
#> [560] "analysis of data from the department of health of northern ireland for th…"
#> [561] "this chapter investigates the policy and practice of established u s labo…"
#> [562] "we are living at a time when events move more rapidly than in the past it…"
#> [563] "introduction hypoparathyroidism which is a common complication following …"
#> [564] "abstracthepatoma derived growth factor hdgf was first purified as a growt…"
#> [565] "background macrophages are implicated in atherosclerotic plaque instabili…"
#> [566] "the present study aims to a explain challenges experienced by midwives b …"
#> [567] "objectivethe monoaminergic stabiliser osu6162 has in previous studies sho…"
#> [568] "the paper studies the impact of regional climate changes that occurred in…"
#> [569] "to estimate the prevalence of antibodies to borrelia burgdorferi in the m…"
#> [570] "We investigated synanthropic small mammals in the Ethiopian Highlands as …"
#> [571] "african great apes are susceptible to infections with several species of …"
#> [572] "the long barrow lies one mile north west of the village close to the nort…"
#> [573] "the united states has negotiated bilateral open skies agreements to dereg…"
#> [574] "this article describes barriers to successful patient outcomes in home he…"
#> [575] "results of a survey for ixodid ticks and or serum antibodies to borrelia …"
#> [576] "in order to determine the role of coyotes in the epidemiology of granuloc…"
#> [577] "summaryfive patients with mild bleeding tendencies characterized by defec…"
#> [578] "on 2 may 2016 robert penner canadian national residing in nepal with a wo…"
#> [579] "thirteen strains of streptococcus equi subsp ruminatorum from free rangin…"
#> [580] "abstract background infection with severe acute respiratory syndrome coro…"
#> [581] "the molecule of the title compound c25h24o obtained by acid catalysed 1 3…"
#> [582] "pelargoniums represent a major floricultural crop in international market…"
#> [583] "abstractprotease activated receptors pars are stimulated by proteolytic c…"
#> [584] "abstract human caliciviruses hucvs cause waterborne outbreaks of gastroen…"
#> [585] "background aims there are no established methods for patients with hemipa…"
#> [586] "seismic attributes derived quantities such as p wave and s wave impedance…"
#> [587] "rainfall erosivity is a measure for the erosive force of rainfall rainfal…"
#> [588] "introduction there are no randomized studies that assess the role of risk…"
#> [589] "abstractcell free dna cfdna is attractive for many applications including…"
#> [590] "information communications technologies icts are one of the major areas o…"
#> [591] "abstractwe argue that a method developed by ångström ann phys chem vol 11…"
#> [592] "several potential mammalian reservoirs of sylvatic species of trichinella…"
#> [593] "abstractwe present the analysis of global sympagic primary production pp …"
#> [594] "two menaquinone deficient and one aromatic deficient mutants of staphyloc…"
#> [595] "harbor seals phoca vitulina richardsi were captured in the coastal region…"
#> [596] "in 1909 president theodore roosevelt personally wrote 200 notes in longha…"
#> [597] "this review discusses the use of optogenetics in the spinal cord and peri…"
#> [598] "carotenoids are organic pigments commonly synthesized by plants algae and…"
#> [599] "the aim of this work is to describe the design of an innovative test rig …"
#> [600] "introduction despite an increase in the number of cadaver donors and over…"