Get a document term matrix from abstracts

Get a document term matrix from abstracts

Usage

get_dtm(abstracts, vocabulary, tf_idf = TRUE)

Arguments

abstracts

Could be an object of class abstracts (created via get_abstracts function) or a charactervector with abstracts

vocabulary

A vocabulary created with the get_vocabulary function, from an abstracts object

tf_idf

logical. If it is true returns dtm weighted with tf idf. Otherwise return binary dtm matrix

Value

Returns a document term matrix. Each row is a document, each term is a matrix from vocabulary

Examples

abstracts <- get_abstracts(lacsSample)
vocabulary <- get_vocabulary(abstracts, term_count = 10)
get_dtm(abstracts, vocabulary, tf_idf = FALSE)
#> 600 x 21 sparse Matrix of class "dgCMatrix"
#>   [[ suppressing 21 column names ‘determine’, ‘infection’, ‘present’ ... ]]
#>                                              
#> 1   . . . . . . . . . 2 . 5 1 . . . . . 1 . .
#> 2   . . . . . . . . . 3 . 2 . . . 1 1 . . 1 2
#> 3   . 3 1 . . . . . . . . . . . . . . . . . .
#> 4   . 3 1 . . . . 2 . . . . 1 . . . . . 1 . .
#> 5   . . . . . . . . . . . . . . 1 . . . 3 1 .
#> 6   . . . . . . . . . 1 . . . . . . . . 3 1 .
#> 7   . . . 2 1 . . . . . . . . 1 . . . . . . .
#> 8   . . 1 . . . . . . . . . . . . . . 1 . . .
#> 9   . 1 . . . . . . . . . . . . . . . . . . .
#> 10  . . . . 6 . . . . . . . . . . . . . . . .
#> 11  . . . . . . 1 . . . . . . . 1 . . . . . 3
#> 12  . . . . . 1 . . . . . . . . . . . . 1 . .
#> 13  . . . . . . . . . . . . . . . . . 1 . . 1
#> 14  . . . . . . . . . . . . . . . . . . . . .
#> 15  . . . . . . . . . . . . . . . . . . . . .
#> 16  . . . 1 . . . . . 5 . . 1 . . . . . . . .
#> 17  . . . . . . 2 . . . 2 . . . 1 . . . . . .
#> 18  . . . . . . . . . . . . . 1 . . . . . . .
#> 19  . . . . . . . . . . . . . . . . . . . . .
#> 20  . . . . . . . . . . . . . . . . . . . . .
#> 21  . 1 . 1 1 . . . . . . . . . . . . . . . .
#> 22  . . . . . 1 . . . . . . 2 . . . . . 1 . .
#> 23  . . . . . . . . 1 . . . . . . . . . 1 . .
#> 24  1 . 1 3 5 . . . . . . . . . . . . . . . .
#> 25  . . . . . . . . . . . . . 3 . . . . 1 . .
#> 26  . . . . . . . . . . . . . . 1 . . . . . .
#> 27  . 1 . . . . . . . . . . . . . . . . . . .
#> 28  . . . . . . . . . . . . 1 . . . . . . . .
#> 29  . . . . . . . . . . . 1 . . . . . . 3 . .
#> 30  . . 1 . . . . . . . . . 1 . . . . . . . 1
#> 31  . . . . . . . . . . . . . . . . . . . . .
#> 32  . . . . . . . . . . . . 1 . . . . . 1 . .
#> 33  . . . . . . . . . . . . . . . . . . 1 . .
#> 34  . . 1 . . . . . . . . . . . . . . . . . .
#> 35  . . . . . . . . . . . 1 . . 1 . . . 1 . .
#> 36  . . . . . . . . . . . . . . . . . . 2 . .
#> 37  . . . . . . . . . . . . . . . . . . . . .
#> 38  . . . . . . . 1 . . . . . . 1 . . . . . .
#> 39  . . . . . . . . . . . . . . . . . . . . .
#> 40  . . . . . . . . . . . . . . 1 . . . . . .
#> 41  . . . . . . 1 . 1 1 . 1 . . . . . 2 1 . .
#> 42  . . . . . . . . . . . . . . . . . . 3 1 .
#> 43  . . . . . . 1 . . . . . . . . . . . . . .
#> 44  1 . . 1 . . . 1 . . . . 1 . 1 . . . 1 . .
#> 45  . . . . . . . . . . . . . . . . . . . . .
#> 46  . . 1 . . . . . . 1 . . . . . . . 1 1 . .
#> 47  . . . 1 . . . . . 2 . . . . . . . . 5 . 1
#> 48  . . . . . . . . . . . . 1 . . . 1 . 1 1 1
#> 49  . . . . . . . . . . . . 1 . . 1 . . . . .
#> 50  . . . . . . . . . . . . . . . . . . 1 . 1
#> 51  . . . . . . 1 . . . . . . . . . . . . . 1
#> 52  1 . 1 . . . . . . . . . . . . . . . 1 . .
#> 53  . . . . . . 1 . . 1 . . . . . . . 1 . . .
#> 54  . . 1 . . . . . . . . . . . . . . . . . 1
#> 55  . 1 . . . . . . . . . . . . . . . . . . .
#> 56  . . . . . . . . . . . . . . . . . . . . .
#> 57  . . . . . . . . . . . . . . . . 1 . 3 1 1
#> 58  . . . . . . . . . . . . 1 2 1 . . . 1 . 1
#> 59  . . 1 . . . . . . . . . . . . . . . . . .
#> 60  . 4 . 1 . 1 . 3 . . . . . 6 . . . . . . .
#> 61  . . . . . . . . . . . . 1 . . . . . 1 . 1
#> 62  . . . . . . . . . . . . . . . . . . . 1 1
#> 63  . 2 . . . . . . . . . . . . . . . . . . .
#> 64  . . . . . . 1 . . . . . . . . . . . . . .
#> 65  . . . . . . . . . . . . . . . . . . . . .
#> 66  . . . . . . . . . . . . 1 . . 1 . . . . .
#> 67  . . . . . . 1 . . . . . . . . 1 . . . . 1
#> 68  . . . . . . . . . . . . . . . . . . . . .
#> 69  . . . 1 . . . . . 2 . 1 . . . . . 1 . . 1
#> 70  . . . . . . . . . . . . . . . . . . 2 . .
#> 71  . . . . . . . . . 1 . . . . . . . . . . .
#> 72  . . . . . . . 5 . . . 1 . . . . . . . . .
#> 73  . 1 . . . . . . . . . . 1 . . . . . . . .
#> 74  . . . . . . . . . . 2 . . . . . . . . . .
#> 75  . . 1 . . . . . 1 . . . 1 . . . . . 3 . 1
#> 76  . 1 . . . 1 . . . . . . . 1 . . . . 1 . .
#> 77  . . . . . . . . . . 1 . . . . . . . . . 1
#> 78  . . . . . . . . . . . 1 . . . . . . . . .
#> 79  . . . 5 . 1 . . . . . . 1 1 . . . . . . .
#> 80  . . . . . . . . . . . . 2 . . 1 . . 1 . .
#> 81  1 . . . . . . . . . . . . . . . . . . . .
#> 82  . . 1 . . 2 . . . . . . 1 4 . . . . 1 . .
#> 83  . . . 2 . . . . . . . 1 . . . . . . 1 . .
#> 84  . . . . . . . . . . . . . . . 1 . . . . .
#> 85  . . . . . . . . . . . . . . . . . . . . .
#> 86  . . . . . . . . . . . . . . . . . . . . 1
#> 87  . . . . . . . . . . . . . . . . . . . . .
#> 88  . . . . . . . . . . . . . . . . . 1 . . 1
#> 89  . 1 . . 1 . . . . 1 . . . . 1 1 . . . . .
#> 90  . . . . . . . 2 . . . . . . . . . . 1 . .
#> 91  . . . . . . 1 . . . . . . . . . . 1 1 2 .
#> 92  . . . . . . . . . 1 . 1 . . . . 1 . 1 1 1
#> 93  . . . . . . . . . . . . . . . . . . . . .
#> 94  . 1 . 1 . . . 1 . . . . . . . . . . . . .
#> 95  . . . 1 . . . . . . . 1 1 . . . . . . . .
#> 96  . . . . . . . . . . . 1 . 1 . . . . . . .
#> 97  1 . . . . . 1 . . . . 1 . . . . . 1 . . .
#> 98  . . . . . . . . . . . . . . . . . . . . .
#> 99  . . 1 . . . . . . . . . 1 1 . . . . . . .
#> 100 . . . 1 1 . . 2 . . . . . 1 . . . . . . .
#> 101 1 . . . 1 . . 2 . . . . 1 3 1 . . . 2 . .
#> 102 1 . . . . . . . . . . . . . . . . . 1 . .
#> 103 . . . . . . . . . . . . 1 . . . . . . . .
#> 104 . . . 3 1 1 . . . . . . . 1 1 . . . . . .
#> 105 . . . . . . . . . . . . . 2 . . . 1 . . .
#> 106 . . . . . . . . . . . . 1 . . . . . . . .
#> 107 . . . 1 . . . . . . . . . . . . . . . . .
#> 108 . . . . . . . . . . . . . . . . . . . . .
#> 109 . . . . . . 1 . . . . . . . . . . . . . .
#> 110 . . . . . . . 2 . . . . . . . . . . 1 . .
#> 111 . . 1 . . . . . . . . . . . . . . . 1 . 1
#> 112 . . . . . 3 . . . . . . . . . . . . . . .
#> 113 . . . . . . . . . . . . . . 1 . . . . . .
#> 114 . . 1 . . . . . . . . . 1 . . . . . . . 1
#> 115 . . . . . . . . . 1 . . . . . 1 . . . . .
#> 116 . 5 . . . . . 2 . . . . 1 . . . . . . . .
#> 117 . . . . . . . . . . . . . . . . . . . . .
#> 118 . . . . . . . . . . . . . . . . . . . . .
#> 119 . . . . . . . 1 . . . . . 5 . . . . . . .
#> 120 . . . . . . . . . . . . 1 . . . . . . . .
#> 121 . . . 1 . . . . . . . . . . . . . . . . .
#> 122 . . . . . . . . . . . . 1 . . . . . 1 . .
#> 123 . . . . . . . . 1 . . 1 . . . 1 . . . . .
#> 124 . . . . . . . . . . . . . . . . . 1 . . .
#> 125 . . . . . . . . . . . . . . . . . . . . .
#> 126 . 3 3 . . 3 . 1 . . . . 2 5 1 . . . 2 . .
#> 127 . . 1 1 2 . . . . . . . 1 . . . . . . . .
#> 128 . . . . . . . . . . . . 1 . . . 1 . 1 1 1
#> 129 . . . . . . 1 . . . . . . . 1 . . 1 1 . .
#> 130 . . . . . . . . . . . . . . . . . . 2 . 1
#> 131 . . . 3 1 . . . . . . . 2 . . . . . 2 . .
#> 132 . . . . . . 1 . . . . . 1 . . . . 2 . . .
#> 133 1 . . . 1 . . 1 . . . . 1 . . . 1 . 1 1 1
#> 134 . . . 8 . . 1 . . . . 3 . . . . . . 1 1 1
#> 135 . . . . . . . . . . . . 1 . . . . . . . .
#> 136 . 1 . 1 . . . 1 . . . . 2 1 . 1 . . 1 . .
#> 137 . 1 . . . 2 . 1 . . . . . 2 . . . . . . .
#> 138 . . . . . . . . . . . . . . 1 . . 1 . . 1
#> 139 . . . . . . 1 . . . . 1 . . . . . . . . .
#> 140 . . . . . . 1 . . . . 1 . . . . . . 1 . .
#> 141 . . 1 . 1 . 1 . . . . . . . . . . . . . .
#> 142 . . . . . . . . . . . . 2 3 1 . . . 3 . .
#> 143 . . . 2 . . . . . . . . . . . . . . 1 . 1
#> 144 . 3 . . . . . 4 . . . . 1 . . . . . . . .
#> 145 . . . . . . . . . 2 . . . . 2 1 . . . . .
#> 146 1 1 . . . 3 1 1 . . . . . . . . . . . . .
#> 147 . . 1 . . . . . . . 1 1 . . . . . . . . .
#> 148 . . . . . . 1 . . . . . . . 2 . . . . . .
#> 149 . . . . . . . . . . . . 1 . . . . . . . .
#> 150 . . . . . . . . . . . . . . 1 . . . 1 . .
#> 151 . . . . . . . . . . . . . . 1 . . 1 1 . 1
#> 152 . . . . . . . . . . . 1 . . . . . . . . .
#> 153 . . . . . . . . . . . . . . 1 . . 1 . . .
#> 154 . . . . . . . . . . . . . . . . . . 3 1 .
#> 155 . . . . . . . . . 2 . . . . . . . 1 . . .
#> 156 . . 1 . . . . . . . . . . . . . . . . . .
#> 157 . . . . . . . . . . . . . . . . . . . . .
#> 158 . . . . . . . . . . . . . . . . . 1 . . .
#> 159 . . . 2 . 1 . . . . . . 1 . . . . . . . .
#> 160 . . . . . . . . . . . . . . . . . . . . .
#> 161 . . . . . . . . . . . . . . . 1 . . . . .
#> 162 . . . . . . . . . . . . . . . . . 1 . . .
#> 163 . . . 2 1 . . . . . . . 1 . . . . . 1 . .
#> 164 . . . . . . . 2 . . . . . 2 . . . . . . .
#> 165 . . . . . . . . . . . . . . . . . . . . .
#> 166 . . . . . . . . . . . . . . . . . . . . .
#> 167 . . . . . . . . . 1 . . . . . . . . . 1 .
#> 168 . . . . . . . . . . . . . . . . 1 . . 1 1
#> 169 . . . . . . . . . . . . . . . . . . 1 . .
#> 170 1 . . . . . . . 1 . . . . . . . . . 2 . .
#> 171 . . 1 . . . . . . . . . . . . . . . 1 . .
#> 172 . . . . . . . . . . . . . . . . . . . . .
#> 173 . . . . . . . . . . . . . . . . . . 2 . .
#> 174 . . . . . . . . . . . 1 . . . . . . 1 . .
#> 175 . . . . . . . . . . . . . . . . . . 1 . 1
#> 176 . . . . . . 1 . . . . . . . . . . . . . .
#> 177 . . . . . . . . . . . . . . . . . 1 . . .
#> 178 . 1 . . . . . . . 1 . . . . 1 . . 1 . 1 .
#> 179 . . . . . . . . . . . . . . . . . . . . .
#> 180 . . . . . . . . . . . . . . . . . . . . .
#> 181 . . . . 2 . . . . . . . . . . . . . 1 . .
#> 182 . . . . . . . . . . 1 . . . . . . . . . 1
#> 183 . . . . . . . . . . . . . . . . . . . . .
#> 184 . . . . . . . 3 . . . . 1 . . . . . 1 . .
#> 185 . . . . . . . . . . . . . . . . . 1 1 . .
#> 186 3 . . . . 1 . . . . . . 1 1 . . . . 2 . .
#> 187 . . . . . . . . . . . . . . . . . . . . .
#> 188 . . . . . . . . . . . . . . . . . 1 . . .
#> 189 . . . . . . . . . . . 1 . 1 . . . . . . .
#> 190 . . . . . . . . . . . . . . . . . . . . 1
#> 191 . . . . . . . . . . . . . . . . . . . . .
#> 192 . . . . . . . . . 1 . 1 1 . . . 1 . 1 1 1
#> 193 . . . 2 . . . . 1 . . . . . . . . . 3 1 .
#> 194 . . . . . . . . . . . . . . . . . . . . 1
#> 195 . . . . . . . . . . . . . . . . . 1 . . .
#> 196 . . . . 2 . . . . . . . . . . . . . . . .
#> 197 . . 1 . . . . . . . . . . . . . . . 1 . .
#> 198 . . 1 . . 1 . 4 . . . 1 1 . . . . . 5 . .
#> 199 . . . . . . . . . . . . 1 . . . . 1 . . .
#> 200 . . . . . . . . . . . . . . . . . . . . .
#> 201 . . . . . . . . . . . . . . 1 . . 1 . . 1
#> 202 . . . . . . . . . . . . . . . . . 1 . . .
#> 203 . . . . . . . . . . . . . . . . . . 1 1 .
#> 204 . . . . . . . . . . . . . . . . . . . . .
#> 205 . . . . . 2 . . . . . . 1 2 . . . . 1 . .
#> 206 . . 2 . . . . . . . . 1 . 1 . . . . 1 . .
#> 207 . . . . . . . 1 . . . . . . . . . . 1 . .
#> 208 . . . . . . 1 . . . . . . . . 1 . . . . .
#> 209 . . 1 . . . . . . . . . . . . . . 1 . . 1
#> 210 . 3 . . . 1 . 5 . . . . 3 . 1 . . . . . .
#> 211 . . . . . . . . . . . . . . . . . . . . .
#> 212 . . . . . . . . . . . 1 1 . 1 . . . . . .
#> 213 1 . . 1 . 2 . . . . . . . 6 . . . . 1 . .
#> 214 . . . . . . . . . . . 1 . . . . . . . . .
#> 215 . 1 . . . . . . . . . . 2 1 . . . . . . .
#> 216 . . . . . . . . 1 . . . . . . . . . 3 . .
#> 217 . . 1 . . . . . . 6 . . . . . . . . 2 . .
#> 218 . . . . . . 1 . . . . . . . . . . 1 2 . .
#> 219 2 . . . . . 1 . . . . 1 . . 1 . 1 . 1 1 1
#> 220 . . 1 . . . . . . . . . . . 1 . . . 2 . .
#> 221 1 . . . . . . . 1 1 . . . . . . . . 2 . .
#> 222 . . 1 . . . . . . . . . . 3 . . . . 1 . .
#> 223 . . . . . . . . . . . . . . 2 . . . 3 . .
#> 224 1 . . . . . . . . 2 . . 2 . . . . . . . .
#> 225 . . . . . . . . 1 . . . . . 1 . . . 1 . .
#> 226 . . . . . . . . . . . . . 7 1 . . . . . 1
#> 227 . . 1 . . . . . . . . . 1 . 1 . 1 . 2 1 1
#> 228 . . . . . . . . . . . . . 8 . 1 . . . . .
#> 229 . . . . . . . . . . . . . . . . . . 1 . 1
#> 230 . . . . . . . . . . . . . . . . . . 2 . 1
#> 231 . . . . . . . . . 1 . . . . 1 . . . 1 . .
#> 232 . . . . . . . . . 1 . . . . . . . . . . 4
#> 233 . . . . . . . . . . . . . . . . . 1 . . .
#> 234 1 . . . . . . . 1 . . . . . . . . . 2 . .
#> 235 . . . . . . . . . 2 . . 1 . 2 1 . . 3 . .
#> 236 . . . . . . . . . . . . . . . . . . . . .
#> 237 . . . . . . . . . . . . . . . . . . . . .
#> 238 2 . . . 1 . . . . . . 1 . . . . . . 2 1 2
#> 239 . 1 . 1 . 1 . . . . . . . . . . . . . . .
#> 240 . . . . . . 1 . 1 . . . . . . 1 . . . . .
#> 241 . . . . . . . . . . . 1 . . . 1 . . 2 . .
#> 242 . . . . . . 1 . . . . . . . . . . 1 . . .
#> 243 . . . . . 1 . . . . . . 1 2 . . . . . . .
#> 244 . . . . . . . . . . . . . . . . . . . . .
#> 245 . . . . . . . . 1 . . . . . . . . . 2 . .
#> 246 . . . . . 3 . . . . . . 1 1 . . . . . . .
#> 247 . . . . . . . . . . . . . . . . . . 2 . .
#> 248 1 . . . . . 1 . . . . . . . . . . . . . .
#> 249 . 1 . . . . . . . . . . . . 1 . . . . . .
#> 250 . . 1 . . . . . . 1 . . . . . . . . . . .
#> 251 . . . . . . . . . . . . . . . . . 1 . . .
#> 252 . . . . . . 1 . . . . . . . . . . . . . .
#> 253 3 . . 1 . . . . . . . . . . . . . . 1 . .
#> 254 . . . . . . . . . . . . . . . . . . . . .
#> 255 . . . . . . . . . . . . . . . . . . . . .
#> 256 . . . . . . . . . . . . . . . 1 . . . . .
#> 257 . . . . . . . . . . . . . . . . . . . . .
#> 258 . . . . . . 1 . 1 . . . 2 . . . 1 . 2 1 1
#> 259 . 1 . . . . . . . . . . 1 . . . . . . . .
#> 260 . . . . . . 1 . . . 1 . . . . 2 . . . . .
#> 261 . . . . . . . . . . . . . . . . . . . . .
#> 262 . . . . . . 1 . . . . . . . 1 . . 1 . . .
#> 263 . . . . . . . . . . . . . . . . . . . . .
#> 264 1 1 . 1 . . . 3 . . . . . 3 . . . . . . .
#> 265 1 4 . 1 . . . . . . . . . . . . . . 1 . .
#> 266 2 . . . . . . . 1 . . . . . . . . . 1 . .
#> 267 . 7 . . . . . . . . . . 1 . 1 . . . 2 1 .
#> 268 . . . . . . . . . . . . . . . . . 1 . . .
#> 269 . 1 . 1 . . . . . . . . 1 2 . . . . . . .
#> 270 . . . . . . . . . . . . . . . . . 1 . . .
#> 271 . . . . . . . . . . . 1 . . . . . . . . .
#> 272 . . . . . . . . . . . . 2 . 1 . . . 1 . .
#> 273 . . 1 . . . . . . . . . . . . . . . 1 . .
#> 274 . . . . . . . . . 1 . . . . . . . . 1 . .
#> 275 1 3 . . . . . . . . . . . . 1 . . . 1 . .
#> 276 . . . . . . 1 . . . . . . . . . . . . . .
#> 277 . . . . . . . . . 1 1 . . . . . . . . . 1
#> 278 . 1 . . . 2 . 1 . 1 . . . . . . . . 2 . .
#> 279 . . . 2 3 . . 2 . . . . 1 . . . . . . . .
#> 280 . . . . . . . 2 . . . . . . . . . . . . .
#> 281 . . . . . . . . . . . . . . . 2 . . . . .
#> 282 . . . . . 2 . 1 . . . . . 2 . . . . 1 . .
#> 283 . . . . . . . 2 . . . . . . . . . . 1 . .
#> 284 . . . . . . . . . . . 1 . . . . 1 . 1 1 1
#> 285 . . . . . . . . . . . . . . . . . . . . 1
#> 286 . 2 . . . . . . . . . . . . 1 . . . 1 . .
#> 287 . . . 4 1 . . . . . . . 2 . . . . . . . .
#> 288 . . . . . . . . . . . . . 4 . . . . . . .
#> 289 . . 1 . . . . . . . . . . . . 1 . . . . .
#> 290 . . 1 . . 1 . . . . . . . . . . . . 1 . .
#> 291 . . . . . . . . . . . . . . . . . . . . 1
#> 292 . . . 1 . . . . . . . . . . . . . . . . .
#> 293 . . . . . . . . . . . 1 . . . . . . 1 . 1
#> 294 . . . . . . . . . . 1 . . . 1 . . . 1 . .
#> 295 . . . . . . . . . . . . . . 1 . . . . . 2
#> 296 . . . . . . . . . . . 1 . . . . . . 3 . .
#> 297 . . . 1 . . . . . . . . 1 . . . . . . . .
#> 298 . . . . . . . . . . . . . . . . . . . . .
#> 299 . . 1 . . . . . 1 . . 1 . . . 1 . . 1 . .
#> 300 . 1 . . . . . . . . . . . 1 . . . . . . .
#> 301 . . . . . . . . . . . . . . . . . . . . .
#> 302 . . . . . . . . . . . . . . . . . . . . .
#> 303 . . . . . . . . . . . . . . . . . . . . .
#> 304 . . . . 1 . . . . . . . 1 3 . . . . . . .
#> 305 . . . . . . 2 . . . . . 1 . . . . . 3 . .
#> 306 . . . . . . 1 . . . . . . . . . . . . . .
#> 307 . . . . . . . . . . . . . . . . . . . . .
#> 308 . . . . . . . . 1 . . . . . . . . 2 . . .
#> 309 . . . 3 . . . . . . . . . 1 . . . . . . .
#> 310 . . . . . . . . . . . . . 1 . . . . . . .
#> 311 . . . . . . . . . . . . . . . 1 . . 1 . .
#> 312 1 . . . . . 1 . . . . . 2 . . 1 . . 1 1 .
#> 313 . . . . . . 1 . . 1 . . . . 1 . . . . . 1
#> 314 . . . . . . . . . . . . . . 1 . . . . . 1
#> 315 . . . . . . . . . . . . 1 . . . . . . . .
#> 316 . . . . . . . 8 . . . . . 1 . . . . 1 . .
#> 317 . . 1 . . 1 . . . . . . 1 . . . . . . . .
#> 318 . . . . . . . . . . . 2 . . . 1 . . . . .
#> 319 . . . . . . . . . . . 2 1 . . . . . . . .
#> 320 . . . . . . . . . . 1 . . . . . . . . . .
#> 321 . . . . . . . . . . . 1 . . . . . . 3 1 .
#> 322 . . . . . . . . . . . . . . . . . . . . 1
#> 323 . . . . . . . . . . . . . . . . . . . . 1
#> 324 . . . . . . . . . . . . . . . . . . 5 . .
#> 325 . . . . . . . . . . . 1 . . . . . . 1 . .
#> 326 . . . . . . . . . . . . . . . . . . . . .
#> 327 . . . 3 . . . 4 . 1 . . . 2 1 . . . . . .
#> 328 . 1 . 1 . . . . . . . . . . . . . . . . .
#> 329 . . . . . . . . . . . . . . . . . . . . .
#> 330 . . . . . . . . . . . . . . . . . . . . .
#> 331 1 . . . . . . . . . . 6 1 . . . . . . . 1
#> 332 . . . . . . . . . . . . . . . . . . . . .
#> 333 . . . . . . . . . 1 . . . . 1 . . . . . 1
#> 334 . . . 6 4 . . . . . . . 1 . . . . . 1 . .
#> 335 . . . 1 . . . . . 1 . . 1 . . . 1 . 1 1 1
#> 336 . . . . . . . . . . . . . . . . . . . . 1
#> 337 . . . . . . . . . . . . 1 . . . . . . . .
#> 338 . . . . . . . . . . . . 1 . . . . . . . .
#> 339 . . 1 . . . . . . . . . 1 . . . . . . . .
#> 340 . . . . . . . . . . . . . . . . . . . . .
#> 341 . . . . . . . . . . . . . . . . . . . . 1
#> 342 . . 1 2 1 . . . . . . . . 1 . . . . . . .
#> 343 . . . . . . . . . . . . . . . . . . . . .
#> 344 . . . . . . . . . . . . . . . . 1 . . 1 1
#> 345 1 . . . . . . . . . . . . . . . . 1 . . 1
#> 346 . . . . . . . . . . . . . . . . . . . . .
#> 347 . . . . . . . . . . . 4 . . . 1 . . . . .
#> 348 . . . . . . 2 . . 1 . . . . . 1 . . . . .
#> 349 . . 1 1 . . . . . . . . 1 . . . . . 1 . .
#> 350 . . 1 . . . . . . . . . . . . . . 1 . . 1
#> 351 . . . . . . . . . . . . . . . 2 . . 1 . .
#> 352 1 . . . . . . . . . . . . . . . . . 1 . .
#> 353 . . . . . . . . . . . . 2 1 . . . . . . .
#> 354 . . . . . . . . . . . . 2 . . . . . 1 1 .
#> 355 1 . . . . . . . 2 . . . . . . . . . 2 1 .
#> 356 . . . . . . . . . . . . . . . . . . . . .
#> 357 . . . . . . . . . . . . . . . . . . . . .
#> 358 . 1 . . . 3 . 1 . . . . 1 1 1 . . . 2 . .
#> 359 . . . . . 3 . . . 1 . . 1 1 . . . . 2 . .
#> 360 . . . . . . . . . . . . . . . . . 1 . . .
#> 361 . . . . . . . . . 1 . 2 . . . . . . . . .
#> 362 . . . . . . . . . . . . . . . . . . . . .
#> 363 . . . . . . . . . . . . 1 5 . . . . . . .
#> 364 . . . . . . . . . . . . . . 1 . . . . . .
#> 365 1 . 1 . . . . . . 1 . . . 1 . . . . . . .
#> 366 . . . 1 . . . . 1 1 . . . . . . . . . . .
#> 367 . . . . . . . . . . . . . . . . . 1 2 . .
#> 368 . . . . . . . . . 1 . . . . 1 . . . 1 . .
#> 369 . . 2 . . . . . . . . . . 2 . . . . . . .
#> 370 . . . . . . . . . 1 . . . . . . . 1 . . .
#> 371 1 . . . . . . . . . . . . . . . 1 . . 1 1
#> 372 . . . . . . . . . . . . . . . . . . . . .
#> 373 . . . . . . . . . . . . 1 . . . . . . . .
#> 374 . . . . . . . . . . . . . . . . . . 1 . .
#> 375 1 . . . . . . . 2 . . . 1 . . . . . 1 . .
#> 376 . . . . . . . . . . . . . . . . . 1 2 . .
#> 377 . . . . . . . . . . . 1 . 3 . . . . . . .
#> 378 . . . . . . . . . . . . . . . . . . . . .
#> 379 . . 1 . . . . . . . . . 1 . . . 1 . 1 1 1
#> 380 . . . . . . . . . . . . . . . . . . 1 . .
#> 381 . . . . . . 3 . . 2 . . 1 . . . . . 3 . .
#> 382 . . . . . . . . . . . . 1 . . . . . . . .
#> 383 . 1 . . . . . 1 . . . . . . . . . . . . .
#> 384 . . 1 . . . . . . . . . . 1 . . . . 1 . .
#> 385 . . . . . . . . . . . . . . . . . . . . 1
#> 386 . . . . . 2 . 2 . . . . . 2 . . . . 1 . .
#> 387 . . . . . . . . . . . . . . . . . . . . .
#> 388 1 2 . . 4 . . . . . . . . . . . . . . . .
#> 389 . . . . . . . . . . . . . . . . . . . . .
#> 390 1 . . . . . . . . . . . . . . . . . . . .
#> 391 . . 1 . . . . . . . . . . . . . . . . . .
#> 392 . 3 . . . . . . . . . . . 4 . . . . . . .
#> 393 . . . . . . . . . . . . . . . . . . . . .
#> 394 . 1 . 1 . 1 . 1 . . . . 1 1 . . . . . . .
#> 395 . . . . . . . . . . . . . . . . . . . . .
#> 396 . . . . . 1 . 1 . . . . . . . . . . . . .
#> 397 . . . . . . 2 . . . . . 1 . 2 . . . 1 . .
#> 398 . . 1 . . . 1 . . 1 . 1 . . . 1 . . . 1 .
#> 399 . . . . . . . . . . . . . . . . . . . . .
#> 400 . . . . . . . . . . . . . 1 . . . . . . .
#> 401 . . . . . . 1 . . . . . . . 1 . . . . . .
#> 402 . . 1 . . . . . . . . . . . . . . . . . 1
#> 403 . . . . . . . . . . . . . . 1 . . . . . .
#> 404 1 . . . . . . . . . . . . 5 . . . . 1 . .
#> 405 . . 1 . . . . . . . . . . . . . . . . . .
#> 406 . . . . . . . . . . . . . . 1 . . . . . .
#> 407 . . . . . . . . . . . . . . . . . . . . .
#> 408 . . . . . . . . . . . . . . . . . . . . 1
#> 409 1 1 . 4 . . . 1 . . . . . 1 . . . . 1 . .
#> 410 . 1 . . . . . . . . . . . 1 . . . . 2 . .
#> 411 . . . 1 1 . . . . . . . 1 . . . . . . . .
#> 412 . . . . . . . . . . . . . . . . . . . . .
#> 413 . . 1 . . . . . . . . 1 . . . . . . 1 . .
#> 414 . . . . . . 1 . . . . . 1 . . . . . . . .
#> 415 . . . . . . 5 . . . . . . . . . . . 1 . .
#> 416 . . . . . . . . . . . . . . 2 1 . . 1 . .
#> 417 1 . . . . . . . . 1 . . . . . . . . 4 . .
#> 418 . . . . . . . . . . . . 1 . . . . . . . .
#> 419 . . 1 . . . . . . . . . 1 . . . . . . . 1
#> 420 . . . . . . . . . . . . . . . . . 1 . . .
#> 421 . . . . . . 1 . . . 2 . . . 1 . . . . . .
#> 422 . . . . . . . . . . . 1 . . . . . 3 1 . .
#> 423 . . . . . . . . . . . . . . . . . . . . .
#> 424 . . . . . . . . . . . . . . . . . . 1 . .
#> 425 . . . . . . . . . . . . . . . . . . . . .
#> 426 . . . . . . . . . . . . . . . . . . . . .
#> 427 . . . . . . . . . . . . . . . . . . . . .
#> 428 . . . . . . . . . . . . . . . 1 . . . . .
#> 429 . . . . . . . . . . . . . . . . . . 2 . .
#> 430 . . . . . . . . . . 2 . . . . . . . 1 . .
#> 431 . . . . . . 3 . . . . . . . . . . . . . .
#> 432 . . . . . . 1 . 1 1 . . 1 . 1 . . 2 . . .
#> 433 1 . . . . . . . . . . 1 1 2 . . . . . . .
#> 434 1 . . . . 1 . . . . . . 2 . . . . . . . .
#> 435 . . . . . . . . . . . . . . . . . . . . .
#> 436 . 1 . . . . . . . . . . . . . . . . . . .
#> 437 . 1 . 1 1 . . . . . . . 1 . . . . . 1 . .
#> 438 1 1 . . . . . . . . . . . . . . . . . . .
#> 439 . . . . . . . . . . . . . . . . 1 . . 1 1
#> 440 . . . . . . . . . . . . . . . 1 . . . . 1
#> 441 . . . . . . . . . . . . . . . . . . . . .
#> 442 . . . . . . . . . . 1 . . . . . . . . . 1
#> 443 . . . 2 . . . . . 1 . . . . . . . . . . 1
#> 444 . . . . . . . . . . . . . . . . . . 1 . .
#> 445 . . . . . . . . . . . . . . . . . 2 . . .
#> 446 . . . . . . . . . 2 . . . . . . . 1 . . .
#> 447 . . . . . . . . . . . 1 . . . . . . . . .
#> 448 1 . . . . . . . . 1 . . . . 1 . . . 2 . .
#> 449 . . . . . . . . . . . . . . . . . . . . 1
#> 450 . . . . . . . . . . . . . 1 . . . . . . .
#> 451 . . . . . . . . . . . . . . . . . 1 . . .
#> 452 . 2 1 . . . . 3 . . . . . 2 . . . . 1 . .
#> 453 . . . . . . 1 . . . . . . . . 1 . . . . .
#> 454 . . . . . . . . . . . . . . 1 . . . . . .
#> 455 . . . . . . . . . . . . . . . . . . 3 1 .
#> 456 . . . 1 . . . . . 1 . . . . 1 . . . . . .
#> 457 . . . . . . . . . . . . . . . . . . 1 . .
#> 458 . . 1 . . . . . . . . . 1 . 1 . . . 1 . .
#> 459 . . 1 . . . . . . . . . . . . 3 . . . . .
#> 460 . . . . . . 1 . . . . . . . . . . . . . .
#> 461 . . . . 1 . . . . . . . . . . . . . . . .
#> 462 . . . . . . . . . . . . . . . . . 1 . . .
#> 463 1 3 . . . 2 . 7 . . . . . 1 . . . . . . .
#> 464 . . . . . . 1 . . . . . 1 . . . . . . . .
#> 465 . . 1 1 . 1 . 2 . . . . . 6 2 . . . 1 . .
#> 466 . . . . . . . . . . . . . . . . . . . . .
#> 467 . . . 2 . . . . . . . . . . . . . . 1 . .
#> 468 . 5 . . . . . . . . . . . . 1 . . . . . .
#> 469 . . 1 . . . . . . . . . . . . . . . . . .
#> 470 . . . 1 . . . . . 1 . . . 1 . . . . . . .
#> 471 . . . . 4 . . . . . . . . . . . . . 1 . .
#> 472 . . . 2 2 . . . . . . . 1 . 1 . . . 1 . .
#> 473 . . . . . . . . . . . . . 1 . . . . . . .
#> 474 . . . . . . . . . 1 . . . . . . . . 2 . .
#> 475 . . . 1 . . . . . . . . . . . . . . . . .
#> 476 . . 1 . . . 1 . . . . . . . . . . . . . .
#> 477 . . . . . . . . . . . . 1 . . . . . . . .
#> 478 . . . . . . . . 1 2 . 2 1 . . . . 2 . . .
#> 479 . . . . . . . . 1 . . . . . . 1 . . . . .
#> 480 . . . . . . . . . . . . . . . . . 1 1 . .
#> 481 2 . . . . . . . . . . . 1 1 1 . . . . . .
#> 482 . . . . . . . . . . . . 1 . . . . . . . .
#> 483 1 . . . . . 1 . 1 . . . . . 1 . . . . 1 .
#> 484 1 . . 1 . . . . . 1 . . . . . . . . . . .
#> 485 1 . . . . . . . . . . . . . . . . . . . .
#> 486 . . . . . . . . . . . . . . . . . . . . .
#> 487 . . . . . . . . . . 2 . . . . . . . . . .
#> 488 . . . 3 2 . . 1 . . . . 1 . . . . . . . .
#> 489 . . . . . . . . . . . . . . . . . . . 1 .
#> 490 . . . . . . . . . . . . 2 3 . . . . . . .
#> 491 . . . . . . . . . . . . . . . . . 1 . . 1
#> 492 . 1 . 2 1 3 . . . . . . . 1 1 . . . . . .
#> 493 . 1 . . . . . . . . . . . . . . . . . . .
#> 494 . . . . . . . . . . . . . . . . . . . . .
#> 495 . . . . . . . . . . . . . . . . . . . . 1
#> 496 . . 2 . . . . . . . . 8 . . . . 1 . 1 1 1
#> 497 . . . . . . 1 . . 1 . . . . . . . . . . .
#> 498 . . . . . . . . . . . . . . . . . . . . .
#> 499 . 1 . 2 . . . . . . . . . 2 . . . . 2 . .
#> 500 . . . . . . . . . . . . . . . . . . . . .
#> 501 . . . . . . . . . . . . 1 . . 1 . . . . 1
#> 502 . . . . . . . . . . . . . . . . . . 1 . 1
#> 503 1 . . . . . . . . . . . 1 . . . . . 3 . .
#> 504 1 . . . . . . . . 4 . . 1 . . . . . 1 . .
#> 505 . 2 . 1 . . . 1 . . . . . 1 . . . . . . .
#> 506 . . . . . . . . . . . . . . . . . . . . .
#> 507 . . . . . . 3 . . 1 . . . . . . . 1 . . 1
#> 508 . . . . . . . . . . . . . . 2 . . . . . .
#> 509 . . . . . . . . 1 . . 1 . . 1 . . 1 . . .
#> 510 . . . . . . . . . . . . . . . . . . 2 1 .
#> 511 . . . . . . . . . . . . . . . 1 . . . . .
#> 512 . 2 . . . 1 . . . . . . 1 1 . . . . . . .
#> 513 . . . . . . . . . . . . . . . . . . . . .
#> 514 . . . . . . . . . . . . . . . . . . . . .
#> 515 . . 1 . . . . . . . . . . . . . . . . . .
#> 516 . . . 1 1 3 . . . . . . . . . . . . 1 . .
#> 517 . . 1 . . . . . . . . . . . . . . 1 . . .
#> 518 . . . . . . . . 1 . . . . . . . . . . . .
#> 519 . . . . . . . . . . . . . 1 . . . . 1 . .
#> 520 . . . . . . . . . . . . . . 1 . . . . . .
#> 521 . . . . . 1 . . . . . . . . 1 . . . . . .
#> 522 . . . . 1 . . 1 . . . . . . . . . . . . .
#> 523 1 . . . . . . . . . . . . . . . . . 2 . .
#> 524 . . . . . . . . . . . . 2 3 . . . . . . .
#> 525 . . . . . . . . . 1 . . . . 1 . . 1 3 1 .
#> 526 . . . 2 . . . . . . . . . . . . . . . 1 .
#> 527 . . . . . . . . . . . . . . . . . . . . .
#> 528 . . . . . . 2 . . 1 . 1 . . 1 . . . . . .
#> 529 . . . . . . . . . . . . . . . . . . . . .
#> 530 . . . . . 2 . . . . . . . 1 2 . . . . . .
#> 531 1 . . . . . 1 . . . . . . . 1 . . . 1 1 .
#> 532 . . . . . . . 1 . . . . . . . . . . . . .
#> 533 . . . . . . . . . . . . . 2 1 . . . 3 . .
#> 534 . . 1 . . . . . . . . . . . . . . . 1 . .
#> 535 . . . . . 1 . . . . . . . . . . . . . . .
#> 536 . . . . . . . . . 3 . . . . . 2 . . . . .
#> 537 . . . . . . . . . . . . . . . . . . . . .
#> 538 . . . . . . . . . . . . 1 . . . . . . . 2
#> 539 . . . . . . . . . . . . . . . . . 1 1 . .
#> 540 . . . . . . . . . 1 . . . . . . . 2 1 . .
#> 541 . . . . . . . . . . . . . . 2 2 . . 1 . .
#> 542 . . 1 . . . . . . . . . . . . 1 . . . . 1
#> 543 . . . . . . . . . . . . . . . . . . . . .
#> 544 . . 1 . . . 1 . . . . 1 2 2 . . . 1 . . .
#> 545 . . . . . . 1 . . . . . . . . . . 1 . . .
#> 546 . . . . . . . . . . . . 1 1 . . . . . . .
#> 547 . . . . . . . . . . . . . . . . . . . . .
#> 548 . . . . . . . . . 1 . . . . . . . . 1 . .
#> 549 . 1 1 1 . . . . . . . . 1 3 . . . . . . .
#> 550 . . . . . . . . . 1 . . . . . . . 1 . . .
#> 551 . . . . . . . 1 . . . . . . . . . . . . .
#> 552 . . 1 . . 3 . 1 . . . . . . . . . . 2 . .
#> 553 . . . . . . . . . . . . . . . . 1 1 . 1 1
#> 554 . . . . . . 1 . 2 1 . . . . 1 . . . 3 . .
#> 555 . . 1 . . . 1 . 1 2 . 1 . . 1 . . 1 3 . .
#> 556 . . 1 . . . . . . . . . 1 . . . . . 1 . .
#> 557 . . . . . . . . . . . . . . . . . . . . .
#> 558 1 . . . . . . . . . . . . . . . . . 1 . .
#> 559 . . . . . . . . . . . . 1 . 1 . . . . . .
#> 560 . . . . . . . . . . . . . . . . . . . . .
#> 561 . . . . . . . . . . 1 . . . . . . . . . .
#> 562 . . . . . . . . . . . . . . . 2 . . . . .
#> 563 . . . . . . . . . . . . . . 1 . . . . . .
#> 564 . . . . . . . . . . . . . . . . . . . . .
#> 565 . . . . . . . . . . . . . . . . . . . 1 .
#> 566 . . 1 . . . . . . . . . . . . . . . 4 . .
#> 567 . . . . . . . . . . . 1 2 . . . . . 1 . .
#> 568 2 . . . . . . . . 1 . 1 . . . . . 1 . . .
#> 569 . . . 2 1 . . 2 . . . . . 1 . . . . . . .
#> 570 . . . . . . . . . . . . 2 . . . . . . . .
#> 571 . 4 1 5 . . . 1 . . . . 1 4 . . . . 1 . .
#> 572 . . . . . . . . . . . . . . . . . . . . .
#> 573 . . . . . . . . . . . . . . . . . . . . .
#> 574 . . . . . . . . . . . . . . . 1 . . . . .
#> 575 . . . 1 1 . . . . . . . 1 3 . . . . . . .
#> 576 1 . . 5 1 . . . . . . . . . . . . . . . .
#> 577 . . . . . . . . . . . . . . . . . . . . .
#> 578 1 . 1 . . . . . . . . 2 . . . 1 . . . . .
#> 579 . . . . . . . . . . . . . 2 . . . . 1 . .
#> 580 . 4 . 1 . . . . . . . . . . . . 1 . . 1 1
#> 581 . . . . . . . . . . . . . . . . . . . . .
#> 582 . . . . . . . . . 1 . . . . . . . . . . .
#> 583 . . . . . . . . . . . . . . . . . . . . .
#> 584 . . . 3 . . . . . . . . 1 . . . . . . . 1
#> 585 . . . . . . . . . . . . . . . . . . 2 1 .
#> 586 . . . . . . . . . . . . . . . . . . . . .
#> 587 . . . . . . . . . 1 . . . . . . . . . . .
#> 588 . . . . . . . . . 1 . 2 . . . . . . 2 . .
#> 589 . . 1 . . . . . . . . . . . . . . . . . .
#> 590 . . . . . . . . . 1 3 . . . . . . . . . .
#> 591 1 . . . . . . . . . . . . . . . . . . . .
#> 592 . 1 . . . . . . . . . . . 1 . . . . . . .
#> 593 . . 1 . . . . . . . . . 1 . . . . . . . .
#> 594 . . . . . . . . . . . . . . . . . . . . .
#> 595 . . . . . . . 3 . . . . . . . . . . 2 . .
#> 596 . . 2 . . . 1 . 1 . . . . . . . . . . . .
#> 597 . . . . . . . . . . . . . . . . . . . . .
#> 598 . . . . . . . . . . . . . . . . . . . . .
#> 599 . . . . . . . . . . . . . . . . . . 1 . .
#> 600 . . 1 . 1 . . . . . . . . . . . . . . . .