Fit a plus model to an abstracts object

Fit a plus model to an abstracts object

Usage

fit_plus(
  abstracts,
  vocabulary = NULL,
  tf_idf = TRUE,
  alpha = 1,
  sample_use_time = 30,
  learning_rate = 1,
  qq = 0.1
)

Arguments

abstracts

An abstracts object.

vocabulary

A ext2vec_vocabulary object from a specific corpus to create a document - term matrix.

tf_idf

Logical. Default is true, use a tf_idf in the document - term matrix S

alpha

Plus parameter. The elastic net mixing parameter, with \(0\le\alpha\le 1\)

sample_use_time

Plus parameter. Number of use time.

learning_rate

Plus parameter. Learning rate. Default is 1.

qq

Plus parameter. quantile threshold

Value

Return a plus object.(Positive and unlabeled Learning from Unbalanced cases and Sparse structures, PLUS). This is a cross-validated glmnet logistics regression optimized for positive and unlabeled learning. This classification is applied to a dtm matrix created with the vocabulary and the abstracts. The independent variables are the possitive and unknown classes of the provided abstract object.

Examples

abstracts <- get_abstracts(lacsSample)
v <- get_vocabulary(abstracts, term_count = 2)
s <- sample(seq(nrow(abstracts)), 65, replace = FALSE)
train <- abstracts[s, ]
model <- fit_plus(abstracts = train, vocabulary = v)
#> Error in plus::plus(x = x, y = y, alpha = alpha, sample_use_time = sample_use_time,     learning_rate = learning_rate, qq = qq): unused argument (alpha = alpha)
model
#> Error in eval(expr, envir, enclos): object 'model' not found