Predicted classification with a lacs model

Predicted classification with a lacs model

Usage

predict(object, ...)

# S3 method for class 'lacs'
predict(object, ..., abstracts, type = "class", plus = TRUE)

Arguments

object

An object of lacs class to predict the text classification

...

additional arguments

abstracts

An object of class abstracts. Is a data.frame with abstracts

type

The type of the predicted response given a lacs model. Could be the class (Binary output) or the probability given the plus model inside the lacs object

plus

If TRUE use the cutoff threshold to predict the class from the plus object. If FALSE use the cv.glmnet object without cutoff threshold.

Value

A vector of response. If the type is set as class returns a logical vector. If the type is response return a vector of probabilities.

Examples

model <- lacs(lacsSample[1:200, ])
lacsSample <- get_abstracts(lacsSample)
predict(model, abstracts = lacsSample )
#>   [1] FALSE FALSE FALSE TRUE  FALSE FALSE TRUE  FALSE TRUE  FALSE FALSE TRUE 
#>  [13] FALSE TRUE  FALSE FALSE FALSE FALSE FALSE FALSE FALSE TRUE  FALSE TRUE 
#>  [25] FALSE FALSE TRUE  FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [37] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [49] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE TRUE  TRUE 
#>  [61] FALSE FALSE TRUE  FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE TRUE 
#>  [73] FALSE FALSE FALSE FALSE FALSE FALSE TRUE  FALSE TRUE  FALSE FALSE FALSE
#>  [85] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [97] FALSE FALSE TRUE  FALSE TRUE  FALSE FALSE TRUE  TRUE  TRUE  FALSE FALSE
#> [109] FALSE TRUE  FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [121] FALSE FALSE FALSE FALSE FALSE TRUE  TRUE  FALSE FALSE FALSE FALSE FALSE
#> [133] FALSE FALSE TRUE  TRUE  FALSE FALSE FALSE FALSE FALSE FALSE FALSE TRUE 
#> [145] FALSE TRUE  FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [157] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [169] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [181] FALSE FALSE FALSE TRUE  TRUE  FALSE FALSE FALSE TRUE  FALSE FALSE FALSE
#> [193] FALSE FALSE FALSE FALSE TRUE  TRUE  FALSE FALSE FALSE FALSE FALSE FALSE
#> [205] TRUE  FALSE TRUE  FALSE FALSE FALSE FALSE FALSE TRUE  FALSE FALSE FALSE
#> [217] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [229] FALSE FALSE FALSE TRUE  FALSE FALSE FALSE FALSE FALSE FALSE TRUE  FALSE
#> [241] FALSE FALSE TRUE  FALSE FALSE TRUE  TRUE  FALSE TRUE  FALSE FALSE FALSE
#> [253] FALSE FALSE FALSE FALSE FALSE FALSE TRUE  FALSE FALSE FALSE FALSE FALSE
#> [265] TRUE  FALSE FALSE FALSE FALSE FALSE FALSE FALSE TRUE  FALSE FALSE FALSE
#> [277] FALSE FALSE TRUE  FALSE FALSE FALSE TRUE  FALSE FALSE FALSE FALSE FALSE
#> [289] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [301] FALSE FALSE FALSE TRUE  FALSE FALSE FALSE FALSE TRUE  FALSE FALSE FALSE
#> [313] FALSE FALSE FALSE FALSE TRUE  FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [325] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE TRUE  FALSE FALSE
#> [337] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [349] FALSE TRUE  FALSE TRUE  FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [361] FALSE FALSE FALSE TRUE  FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [373] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE TRUE  FALSE
#> [385] FALSE TRUE  FALSE FALSE FALSE TRUE  TRUE  FALSE FALSE FALSE FALSE FALSE
#> [397] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE TRUE  FALSE FALSE FALSE
#> [409] FALSE FALSE TRUE  FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [421] FALSE FALSE FALSE TRUE  FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [433] FALSE FALSE FALSE FALSE TRUE  TRUE  FALSE FALSE FALSE FALSE FALSE FALSE
#> [445] FALSE FALSE FALSE FALSE FALSE FALSE FALSE TRUE  FALSE FALSE FALSE FALSE
#> [457] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE TRUE 
#> [469] FALSE TRUE  FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [481] FALSE FALSE FALSE FALSE FALSE FALSE FALSE TRUE  FALSE FALSE FALSE TRUE 
#> [493] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [505] FALSE FALSE FALSE FALSE FALSE FALSE FALSE TRUE  FALSE FALSE FALSE FALSE
#> [517] TRUE  TRUE  FALSE FALSE FALSE TRUE  FALSE TRUE  FALSE FALSE FALSE FALSE
#> [529] FALSE FALSE FALSE TRUE  FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [541] FALSE FALSE FALSE FALSE FALSE FALSE TRUE  FALSE FALSE FALSE FALSE TRUE 
#> [553] FALSE FALSE TRUE  FALSE TRUE  FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [565] FALSE TRUE  FALSE FALSE TRUE  FALSE TRUE  FALSE FALSE FALSE FALSE TRUE 
#> [577] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [589] FALSE FALSE FALSE FALSE TRUE  FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> Levels: FALSE TRUE
predict(model, abstracts = lacsSample,  plus = FALSE )
#>   [1] 0 0 0 1 0 0 0 0 1 0 0 1 0 1 0 0 0 0 0 0 0 1 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0
#>  [38] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 1 0 0 1 0 0 0 0 0 0 0 0 1 0 0
#>  [75] 0 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 1 0 0 1 0 1 0 0 0 1 0
#> [112] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 1 1 0 0 0 0 0 0 0 1 1 0 0 0 0 0 0 0 1 0 0 0 0
#> [149] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 1 0
#> [186] 0 0 0 0 0 0 0 0 0 0 0 1 1 0 0 0 0 0 0 1 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> [223] 0 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0 1 0 0 0 1 0 0 1 1 0 1 0 0 0 0 0 0 0 0 0 1
#> [260] 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0 1 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0
#> [297] 0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> [334] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> [371] 0 0 0 0 0 0 0 0 0 0 0 0 1 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> [408] 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0 1 1 0 0 0 0 0 0
#> [445] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 1 0 1 0 0 0 0 0 0 0 0 0 0 0
#> [482] 0 0 0 0 0 0 1 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0
#> [519] 0 0 0 1 0 1 0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 1 0 0 0 0 1 0 0 0
#> [556] 0 0 0 0 0 0 0 0 0 0 0 0 0 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0
#> [593] 0 0 0 0 0 0 0 0
predict(model, abstracts = lacsSample, type = "response")
#>      lambda.min
#> 1   0.032180185
#> 2   0.014170953
#> 3   0.032180185
#> 4   0.999999999
#> 5   0.032180185
#> 6   0.032180185
#> 7   0.961699942
#> 8   0.032180185
#> 9   0.902944100
#> 10  0.032180185
#> 11  0.001160122
#> 12  0.999951721
#> 13  0.006176043
#> 14  0.999998994
#> 15  0.032180185
#> 16  0.032180185
#> 17  0.032180185
#> 18  0.032180185
#> 19  0.032180185
#> 20  0.032180185
#> 21  0.032180185
#> 22  0.999999960
#> 23  0.032180185
#> 24  0.998991445
#> 25  0.032180185
#> 26  0.032180185
#> 27  0.890651596
#> 28  0.032180185
#> 29  0.032180185
#> 30  0.006176043
#> 31  0.032180185
#> 32  0.032180185
#> 33  0.032180185
#> 34  0.032180185
#> 35  0.032180185
#> 36  0.032180185
#> 37  0.032180185
#> 38  0.032180185
#> 39  0.032180185
#> 40  0.032180185
#> 41  0.032180185
#> 42  0.032180185
#> 43  0.032180185
#> 44  0.032180185
#> 45  0.032180185
#> 46  0.504856234
#> 47  0.010752355
#> 48  0.016715371
#> 49  0.032180185
#> 50  0.010752355
#> 51  0.014170953
#> 52  0.195375068
#> 53  0.032180185
#> 54  0.032180185
#> 55  0.032180185
#> 56  0.032180185
#> 57  0.022392037
#> 58  0.001160122
#> 59  0.998958100
#> 60  0.999999328
#> 61  0.016715371
#> 62  0.006176043
#> 63  0.902944100
#> 64  0.032180185
#> 65  0.032180185
#> 66  0.032180185
#> 67  0.001160122
#> 68  0.032180185
#> 69  0.014170953
#> 70  0.032180185
#> 71  0.032180185
#> 72  0.999998498
#> 73  0.032180185
#> 74  0.032180185
#> 75  0.006176043
#> 76  0.032180185
#> 77  0.016715371
#> 78  0.032180185
#> 79  1.000000000
#> 80  0.032180185
#> 81  0.827341912
#> 82  0.032180185
#> 83  0.032180185
#> 84  0.032180185
#> 85  0.032180185
#> 86  0.006176043
#> 87  0.032180185
#> 88  0.014170953
#> 89  0.240476735
#> 90  0.691973391
#> 91  0.032180185
#> 92  0.020470067
#> 93  0.032180185
#> 94  0.032180185
#> 95  0.683269869
#> 96  0.032180185
#> 97  0.032180185
#> 98  0.032180185
#> 99  0.875120387
#> 100 0.700542565
#> 101 0.996137008
#> 102 0.032180185
#> 103 0.032180185
#> 104 1.000000000
#> 105 0.970047502
#> 106 0.999789912
#> 107 0.032180185
#> 108 0.032180185
#> 109 0.032180185
#> 110 0.999995072
#> 111 0.014170953
#> 112 0.032180185
#> 113 0.032180185
#> 114 0.006176043
#> 115 0.032180185
#> 116 0.228264505
#> 117 0.032180185
#> 118 0.032180185
#> 119 0.342282198
#> 120 0.032180185
#> 121 0.032180185
#> 122 0.032180185
#> 123 0.032180185
#> 124 0.032180185
#> 125 0.032180185
#> 126 0.999999943
#> 127 0.999999999
#> 128 0.016715371
#> 129 0.032180185
#> 130 0.039144900
#> 131 0.032180185
#> 132 0.032180185
#> 133 0.016715371
#> 134 0.010752355
#> 135 0.999990447
#> 136 1.000000000
#> 137 0.032180185
#> 138 0.006176043
#> 139 0.032180185
#> 140 0.032180185
#> 141 0.032180185
#> 142 0.032180185
#> 143 0.010752355
#> 144 0.999998965
#> 145 0.032180185
#> 146 0.926471636
#> 147 0.032180185
#> 148 0.032180185
#> 149 0.032180185
#> 150 0.032180185
#> 151 0.016715371
#> 152 0.032180185
#> 153 0.032180185
#> 154 0.032180185
#> 155 0.032180185
#> 156 0.032180185
#> 157 0.032180185
#> 158 0.032180185
#> 159 0.228264505
#> 160 0.032180185
#> 161 0.032180185
#> 162 0.032180185
#> 163 0.280711001
#> 164 0.032180185
#> 165 0.032180185
#> 166 0.032180185
#> 167 0.032180185
#> 168 0.018655960
#> 169 0.032180185
#> 170 0.032180185
#> 171 0.265902457
#> 172 0.032180185
#> 173 0.032180185
#> 174 0.032180185
#> 175 0.001160122
#> 176 0.032180185
#> 177 0.032180185
#> 178 0.032180185
#> 179 0.032180185
#> 180 0.032180185
#> 181 0.032180185
#> 182 0.014170953
#> 183 0.032180185
#> 184 1.000000000
#> 185 0.999999997
#> 186 0.285285010
#> 187 0.032180185
#> 188 0.032180185
#> 189 0.999999998
#> 190 0.006176043
#> 191 0.032180185
#> 192 0.014170953
#> 193 0.032180185
#> 194 0.001160122
#> 195 0.032180185
#> 196 0.032180185
#> 197 1.000000000
#> 198 0.999889080
#> 199 0.032180185
#> 200 0.032180185
#> 201 0.010752355
#> 202 0.032180185
#> 203 0.032180185
#> 204 0.032180185
#> 205 0.999999960
#> 206 0.032180185
#> 207 0.999999999
#> 208 0.032180185
#> 209 0.001160122
#> 210 0.032180185
#> 211 0.032180185
#> 212 0.032180185
#> 213 0.827341912
#> 214 0.032180185
#> 215 0.032180185
#> 216 0.032180185
#> 217 0.032180185
#> 218 0.032180185
#> 219 0.010752355
#> 220 0.032180185
#> 221 0.032180185
#> 222 0.032180185
#> 223 0.032180185
#> 224 0.032180185
#> 225 0.032180185
#> 226 0.001160122
#> 227 0.014170953
#> 228 0.032180185
#> 229 0.006176043
#> 230 0.006176043
#> 231 0.032180185
#> 232 0.993878394
#> 233 0.032180185
#> 234 0.032180185
#> 235 0.032180185
#> 236 0.032180185
#> 237 0.032180185
#> 238 0.003540485
#> 239 0.999999999
#> 240 0.032180185
#> 241 0.032180185
#> 242 0.032180185
#> 243 1.000000000
#> 244 0.032180185
#> 245 0.032180185
#> 246 1.000000000
#> 247 0.999033621
#> 248 0.032180185
#> 249 0.999615984
#> 250 0.032180185
#> 251 0.032180185
#> 252 0.032180185
#> 253 0.032180185
#> 254 0.032180185
#> 255 0.032180185
#> 256 0.342282198
#> 257 0.032180185
#> 258 0.032180185
#> 259 1.000000000
#> 260 0.032180185
#> 261 0.032180185
#> 262 0.032180185
#> 263 0.032180185
#> 264 0.032180185
#> 265 1.000000000
#> 266 0.032180185
#> 267 0.032180185
#> 268 0.032180185
#> 269 0.032180185
#> 270 0.032180185
#> 271 0.032180185
#> 272 0.032180185
#> 273 0.849538476
#> 274 0.032180185
#> 275 0.032180185
#> 276 0.032180185
#> 277 0.010752355
#> 278 0.032180185
#> 279 0.999989667
#> 280 0.032180185
#> 281 0.032180185
#> 282 0.032180185
#> 283 0.999999787
#> 284 0.014170953
#> 285 0.001160122
#> 286 0.032180185
#> 287 0.032180185
#> 288 0.032180185
#> 289 0.032180185
#> 290 0.032180185
#> 291 0.006176043
#> 292 0.032180185
#> 293 0.006176043
#> 294 0.032180185
#> 295 0.032180185
#> 296 0.032180185
#> 297 0.032180185
#> 298 0.032180185
#> 299 0.032180185
#> 300 0.032180185
#> 301 0.032180185
#> 302 0.032180185
#> 303 0.032180185
#> 304 0.886974277
#> 305 0.032180185
#> 306 0.032180185
#> 307 0.032180185
#> 308 0.032180185
#> 309 0.806916145
#> 310 0.032180185
#> 311 0.032180185
#> 312 0.032180185
#> 313 0.001160122
#> 314 0.001160122
#> 315 0.032180185
#> 316 0.032180185
#> 317 0.999586929
#> 318 0.032180185
#> 319 0.032180185
#> 320 0.032180185
#> 321 0.032180185
#> 322 0.006176043
#> 323 0.001160122
#> 324 0.032180185
#> 325 0.032180185
#> 326 0.032180185
#> 327 0.032180185
#> 328 0.032180185
#> 329 0.032180185
#> 330 0.032180185
#> 331 0.001160122
#> 332 0.032180185
#> 333 0.032180185
#> 334 0.999499060
#> 335 0.010752355
#> 336 0.006176043
#> 337 0.032180185
#> 338 0.032180185
#> 339 0.032180185
#> 340 0.032180185
#> 341 0.001160122
#> 342 0.032180185
#> 343 0.032180185
#> 344 0.010752355
#> 345 0.006176043
#> 346 0.032180185
#> 347 0.032180185
#> 348 0.032180185
#> 349 0.032180185
#> 350 0.994450483
#> 351 0.032180185
#> 352 0.998553996
#> 353 0.032180185
#> 354 0.032180185
#> 355 0.032180185
#> 356 0.032180185
#> 357 0.032180185
#> 358 0.032180185
#> 359 0.032180185
#> 360 0.032180185
#> 361 0.032180185
#> 362 0.032180185
#> 363 0.032180185
#> 364 0.875120387
#> 365 0.032180185
#> 366 0.032180185
#> 367 0.032180185
#> 368 0.032180185
#> 369 0.032180185
#> 370 0.032180185
#> 371 0.016297429
#> 372 0.032180185
#> 373 0.032180185
#> 374 0.032180185
#> 375 0.032180185
#> 376 0.032180185
#> 377 0.032180185
#> 378 0.032180185
#> 379 0.010752355
#> 380 0.032180185
#> 381 0.032180185
#> 382 0.032180185
#> 383 1.000000000
#> 384 0.032180185
#> 385 0.006176043
#> 386 1.000000000
#> 387 0.032180185
#> 388 0.032180185
#> 389 0.032180185
#> 390 0.998553996
#> 391 0.999999964
#> 392 0.032180185
#> 393 0.032180185
#> 394 0.032180185
#> 395 0.032180185
#> 396 0.032180185
#> 397 0.032180185
#> 398 0.032180185
#> 399 0.032180185
#> 400 0.032180185
#> 401 0.032180185
#> 402 0.063405513
#> 403 0.032180185
#> 404 0.285285010
#> 405 0.999999964
#> 406 0.032180185
#> 407 0.032180185
#> 408 0.006176043
#> 409 0.032180185
#> 410 0.032180185
#> 411 0.991881809
#> 412 0.032180185
#> 413 0.032180185
#> 414 0.032180185
#> 415 0.032180185
#> 416 0.032180185
#> 417 0.477524699
#> 418 0.032180185
#> 419 0.001160122
#> 420 0.032180185
#> 421 0.032180185
#> 422 0.032180185
#> 423 0.032180185
#> 424 0.999996787
#> 425 0.032180185
#> 426 0.032180185
#> 427 0.032180185
#> 428 0.032180185
#> 429 0.032180185
#> 430 0.032180185
#> 431 0.032180185
#> 432 0.032180185
#> 433 0.032180185
#> 434 0.032180185
#> 435 0.032180185
#> 436 0.032180185
#> 437 0.991881809
#> 438 0.999615984
#> 439 0.016715371
#> 440 0.010752355
#> 441 0.032180185
#> 442 0.014170953
#> 443 0.032180185
#> 444 0.032180185
#> 445 0.032180185
#> 446 0.032180185
#> 447 0.032180185
#> 448 0.032180185
#> 449 0.001160122
#> 450 0.032180185
#> 451 0.032180185
#> 452 0.999499060
#> 453 0.032180185
#> 454 0.032180185
#> 455 0.032180185
#> 456 0.032180185
#> 457 0.032180185
#> 458 0.032180185
#> 459 0.032180185
#> 460 0.032180185
#> 461 0.032180185
#> 462 0.032180185
#> 463 0.032180185
#> 464 0.032180185
#> 465 0.064924415
#> 466 0.032180185
#> 467 0.032180185
#> 468 0.999615984
#> 469 0.032180185
#> 470 0.999999999
#> 471 0.032180185
#> 472 0.032180185
#> 473 0.032180185
#> 474 0.032180185
#> 475 0.032180185
#> 476 0.032180185
#> 477 0.032180185
#> 478 0.032180185
#> 479 0.032180185
#> 480 0.032180185
#> 481 0.032180185
#> 482 0.032180185
#> 483 0.032180185
#> 484 0.032180185
#> 485 0.032180185
#> 486 0.032180185
#> 487 0.032180185
#> 488 0.999715459
#> 489 0.032180185
#> 490 0.032180185
#> 491 0.006176043
#> 492 0.999999986
#> 493 0.032180185
#> 494 0.032180185
#> 495 0.001160122
#> 496 0.010752355
#> 497 0.032180185
#> 498 0.032180185
#> 499 0.032180185
#> 500 0.032180185
#> 501 0.001160122
#> 502 0.001160122
#> 503 0.477524699
#> 504 0.032180185
#> 505 0.032180185
#> 506 0.032180185
#> 507 0.014170953
#> 508 0.032180185
#> 509 0.032180185
#> 510 0.032180185
#> 511 0.032180185
#> 512 1.000000000
#> 513 0.032180185
#> 514 0.032180185
#> 515 0.032180185
#> 516 0.032180185
#> 517 0.998958100
#> 518 0.992905916
#> 519 0.032180185
#> 520 0.032180185
#> 521 0.032180185
#> 522 0.999998994
#> 523 0.032180185
#> 524 0.999999496
#> 525 0.032180185
#> 526 0.032180185
#> 527 0.032180185
#> 528 0.032180185
#> 529 0.032180185
#> 530 0.032180185
#> 531 0.032180185
#> 532 0.999999999
#> 533 0.126627527
#> 534 0.032180185
#> 535 0.032180185
#> 536 0.032180185
#> 537 0.032180185
#> 538 0.032180185
#> 539 0.032180185
#> 540 0.032180185
#> 541 0.032180185
#> 542 0.006176043
#> 543 0.032180185
#> 544 0.032180185
#> 545 0.032180185
#> 546 0.211246590
#> 547 1.000000000
#> 548 0.032180185
#> 549 0.683269869
#> 550 0.032180185
#> 551 0.032180185
#> 552 1.000000000
#> 553 0.010752355
#> 554 0.032180185
#> 555 0.969008071
#> 556 0.032180185
#> 557 0.993961007
#> 558 0.032180185
#> 559 0.032180185
#> 560 0.032180185
#> 561 0.032180185
#> 562 0.032180185
#> 563 0.032180185
#> 564 0.032180185
#> 565 0.032180185
#> 566 0.998958100
#> 567 0.032180185
#> 568 0.032180185
#> 569 1.000000000
#> 570 0.032180185
#> 571 0.926471636
#> 572 0.032180185
#> 573 0.032180185
#> 574 0.032180185
#> 575 0.565548344
#> 576 0.998553996
#> 577 0.032180185
#> 578 0.032180185
#> 579 0.032180185
#> 580 0.093370578
#> 581 0.032180185
#> 582 0.032180185
#> 583 0.032180185
#> 584 0.006176043
#> 585 0.032180185
#> 586 0.032180185
#> 587 0.032180185
#> 588 0.032180185
#> 589 0.032180185
#> 590 0.032180185
#> 591 0.032180185
#> 592 0.032180185
#> 593 0.998958100
#> 594 0.032180185
#> 595 0.032180185
#> 596 0.032180185
#> 597 0.032180185
#> 598 0.032180185
#> 599 0.032180185
#> 600 0.032180185