Predicted classification with a lacs model

Predicted classification with a lacs model

Usage

predict(object, ...)

# S3 method for class 'lacs'
predict(object, ..., abstracts, type = "class", plus = TRUE)

Arguments

object

An object of lacs class to predict the text classification

...

additional arguments

abstracts

An object of class abstracts. Is a data.frame with abstracts

type

The type of the predicted response given a lacs model. Could be the class (Binary output) or the probability given the plus model inside the lacs object

plus

If TRUE use the cutoff threshold to predict the class from the plus object. If FALSE use the cv.glmnet object without cutoff threshold.

Value

A vector of response. If the type is set as class returns a logical vector. If the type is response return a vector of probabilities.

Examples

model <- lacs(lacsSample[1:200, ])
#> Error in plus::plus(x = x, y = y, alpha = alpha, sample_use_time = sample_use_time,     learning_rate = learning_rate, qq = qq): unused argument (alpha = alpha)
lacsSample <- get_abstracts(lacsSample)
predict(model, abstracts = lacsSample )
#> Error in eval(expr, envir, enclos): object 'model' not found
predict(model, abstracts = lacsSample,  plus = FALSE )
#> Error in eval(expr, envir, enclos): object 'model' not found
predict(model, abstracts = lacsSample, type = "response")
#> Error in eval(expr, envir, enclos): object 'model' not found